% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_methods.R
\name{cal_create_calibration_set}
\alias{cal_create_calibration_set}
\title{Helper function to create a calibration set for ogs5 input parameters}
\usage{
cal_create_calibration_set(...)
}
\arguments{
\item{...}{\emph{vectors} or \emph{lists} of 4 containing parameter locations in
an \emph{ogs5}-object, specifications and its range.}
}
\value{
\emph{data.frame} that can be handed over to \code{\link[=cal_sample_parameters]{cal_sample_parameters()}}
}
\description{
Helper function to create a calibration set for ogs5 input parameters
}
\details{
The 4 elements of each parameter should have the following order:
\enumerate{
\item The location in the \emph{ogs5} object used for calibration as a character
string of the form:
\strong{<file extension>$<main-keyword><sub-keyword>}
\item An additional character string to be pasted before the value in the
ogs5 input file. Can be set to "".
\item The minimum value of the parameter
\item The maximum value of the parameter
}
If a parameter has two values, that should be calibrated simply two identical
parameter specifications should be provided in adjacent rows.
}
\examples{
\dontrun{
calibration_set <- cal_create_calibration_set(
    c("mmp$MEDIUM_PROPERTIES1$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-4, 1.0e-2),
    c("mmp$MEDIUM_PROPERTIES2$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-9, 1.0e-4),
    c("mmp$MEDIUM_PROPERTIES3$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-7, 1.0e-3),
    c("mmp$MEDIUM_PROPERTIES4$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-7, 1.0e-3)
)
}
}
