% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_methods.R
\name{cal_simulation_error}
\alias{cal_simulation_error}
\title{Simulation error for calibration purposes}
\usage{
cal_simulation_error(
  par_df,
  exp_data,
  ogs5_obj,
  outbloc_names,
  ogs_exe,
  log_output = TRUE,
  log_path = NULL,
  run_path = attributes(ogs5_obj)$sim_path,
  target_function,
  ensemble_path = NULL
)
}
\arguments{
\item{par_df}{\emph{tibble} with parameters such as described in \code{\link{cal_bayesOpt}}}

\item{exp_data}{experimental data}

\item{ogs5_obj}{\emph{ogs5}}

\item{outbloc_names}{\emph{character vector} names of the blocs specified in the
*\emph{.out} file that should be used for calibration. The argument will be passed
to the function \code{\link{ogs5_get_output_specific}}.}

\item{ogs_exe}{\emph{character} path to the ogs executable}

\item{log_output}{\emph{logical} Should a log file be written?}

\item{log_path}{\emph{character} path of directory to write log file, default is \code{run_path}.}

\item{run_path}{\emph{character} path where the simulation should be run, default is
\emph{sim_path} from the \code{ogs5_obj}.}

\item{target_function}{\emph{function} specified by the user that should exist in
the global environment and be of the form \verb{f <- function(ogs_obj, exp_data) \{ ... return(sim_error) \}}.}

\item{ensemble_path}{\emph{character} path where ensemble for initial parameters
should be written and run.}
}
\value{
An \emph{ogs5} or \emph{ens} class object depending if one set of parameters or
several were run.
}
\description{
Helper function for \code{\link[=cal_bayesOpt]{cal_bayesOpt()}} that evaluates one or
several parameters values by modifying the \emph{ogs5} input slot, writing input
files and running simulations. The output is retrieved via
\code{\link{ogs5_get_output_specific}} and fed into the user-specified
\code{target_function} to calculate the simulation error.
}
\examples{
r2ogs/examples/bayesOpt_example.R
}
