% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5.R
\name{create_ogs5}
\alias{create_ogs5}
\title{create_ogs5}
\usage{
create_ogs5(
  sim_name = character(NULL),
  sim_id = integer(NULL),
  sim_path = character(NULL)
)
}
\arguments{
\item{sim_name}{Name of the simulation. character}

\item{sim_id}{ID of the simulation. integer}

\item{sim_path}{Path where to export the input files and to run the
simulation. character}
}
\value{
Object of class \emph{og5}
}
\description{
Constructor for the \emph{ogs5} base-class
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ogs5_obj", sim_id = 1L,
                        sim_path = paste0(tmp, "/ogs5_obj"))
}
