% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_additional.R
\name{create_ogs5_additional}
\alias{create_ogs5_additional}
\title{create_ogs5_additional}
\usage{
create_ogs5_additional()
}
\value{
Object of class \emph{ogs5_additional}
}
\description{
Constructor for the \emph{ogs5_additional} class
}
