% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{create_ogs5_gli_polylines}
\alias{create_ogs5_gli_polylines}
\title{create_ogs5_gli_polylines}
\usage{
create_ogs5_gli_polylines()
}
\value{
Object of class \emph{ogs5_gli_polylines}
}
\description{
Constructor for the \emph{ogs5_gli_polylines} class
}
