% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structured_mesh.R
\name{create_structured_mesh_nodes_ele}
\alias{create_structured_mesh_nodes_ele}
\title{create_structured_mesh_nodes_ele}
\usage{
create_structured_mesh_nodes_ele(
  origin = c(0, 0, 0),
  lx = NULL,
  nx = NULL,
  sx = NULL,
  ly = NULL,
  ny = NULL,
  sy = NULL,
  lz = NULL,
  nz = NULL,
  sz = NULL
)
}
\arguments{
\item{origin}{\emph{numeric} vector of c(x,y,z) coordinates.}

\item{lx}{\emph{numeric} length in 'x' direction in 'meter'.}

\item{nx}{\emph{numeric} number of elements in 'x' direction.}

\item{sx}{\emph{numeric} multiplication factor of element size in 'x' direction.}

\item{ly}{\emph{numeric} length in 'y' direction in 'meter'.}

\item{ny}{\emph{numeric} number of elements in 'y' direction.}

\item{sy}{\emph{numeric} multiplication factor of element size in 'y' direction.}

\item{lz}{\emph{numeric} length in 'z' direction in 'meter'.}

\item{nz}{\emph{numeric} multiplication factor of element size in 'z' direction.}

\item{sz}{\emph{numeric} number of elements in 'z' direction.}
}
\value{
\emph{list} of NODES \emph{tibble} and ELEMENTS \emph{tibble}.
}
\description{
Create a structured finite-element mesh in 1D, 2D or 3D.
}
\examples{
mesh_lst <- create_structured_mesh_nodes_ele(lx = 4.7, nx = 94)
}
