% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ens_add_ogs5}
\alias{ens_add_ogs5}
\title{ens_add_ogs5}
\usage{
ens_add_ogs5(x, ogs5_obj)
}
\arguments{
\item{x}{Instanc of class \emph{ens}.}

\item{ogs5_obj}{Instanc of class \emph{ogs5} to be added to \emph{x}.}
}
\value{
Extended instance \emph{x}.
}
\description{
Adds an \emph{ogs5} instance to an existing \emph{ens} instance.
}
