% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ens_get_output}
\alias{ens_get_output}
\title{ens_get_output}
\usage{
ens_get_output(x, type = "all", ...)
}
\arguments{
\item{x}{Instanc of class \emph{ens} containing the simulation objects.}

\item{type}{If 'all' retrieve all produced output, if 'specific'
retrieve output stated in a certain \emph{OUT} bloc of the \emph{ogs5} simulation
objects.}

\item{If}{\emph{type} = 'specific', state names (\emph{list(character)}) of the \emph{OUT}
blocs you wish to retrieve output for.}
}
\value{
Extended \emph{x} with added output.
}
\description{
Wrapper to retrieve output for all included \emph{ogs5} simulation
objects.
}
\examples{
ens1 <- ens_get_output(ens1, type = "all")
}
