% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ens_run}
\alias{ens_run}
\title{ens_run}
\usage{
ens_run(x, ogs_exe, log_output = TRUE, wait = FALSE)
}
\arguments{
\item{x}{Instanc of class \emph{ens} containing the simulation objects.}

\item{ogs_exe}{Path to ogs5 executable. \emph{character}}

\item{log_output}{Write log output? \emph{logical}}

\item{wait}{If \emph{TRUE} R starts the individual simulations runs sequentially,
to say, it waits until a run finishes before starting the next. If \emph{FALSE}
all runs will be started the same time. \emph{logical}}
}
\value{

}
\description{
Wrapper to execute \code{\link[=ogs5_run]{ogs5_run()}} for all included \emph{ogs5} objects
included.
}
\examples{
ens_run(ens1,
        ogs_exe = "../inst/ogs/ogs_5.76",
        log_output = TRUE,
        wait = FALSE)
}
