% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ens_write_inputfiles}
\alias{ens_write_inputfiles}
\title{ens_write_inputfiles}
\usage{
ens_write_inputfiles(x, type = "all")
}
\arguments{
\item{type}{Type of input files to write, see \code{\link[=ogs5_write_inputfiles]{ogs5_write_inputfiles()}}.}

\item{Instanc}{of class \emph{ens} containing the simulation objects.}
}
\description{
Wrapper function to write input files for all \emph{ogs5} simulation
objects belonging to an \emph{ens} instance.
}
