% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundwater_exp}
\alias{groundwater_exp}
\title{Measurements of water levels in 14 wells in the Ashi River basin}
\format{
A data frame with 14 rows and 2 variables:
\describe{
\item{well}{identifier number of the well}
\item{head}{hydraulic head (m)}
...
}
}
\usage{
groundwater_exp
}
\description{
The measurement points consist of hydraulic head (meters) in 14 drinking water
wells and lie in the Ashi River basin in the Northeastern Chinese
province Heilongjiang close to the city Acheng.
}
\keyword{datasets}
