% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_bc.R
\name{input_add_bc_bloc}
\alias{input_add_bc_bloc}
\title{input_add_bc_bloc}
\usage{
input_add_bc_bloc(
  x = list(),
  bc_name = NULL,
  COMP_NAME = NULL,
  CONSTRAINED = NULL,
  COPY_VALUE = NULL,
  DIS_TYPE = NULL,
  DIS_TYPE_CONDITION = NULL,
  EPSILON = NULL,
  EXCAVATION = NULL,
  FCT_TYPE = NULL,
  GEO_TYPE = NULL,
  MSH_TYPE = NULL,
  NO_DIPS_INCREMENT = NULL,
  PCS_TYPE = NULL,
  PRESSURE_AS_HEAD = NULL,
  PRIMARY_VARIABLE = NULL,
  TIME_CONTROLLED_ACTIVE = NULL,
  TIM_TYPE = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{bc_name}{Name of the boundary condition.}

\item{COMP_NAME}{Component name.}

\item{CONSTRAINED}{ogs5  \strong{bc} bloc sub keyword.}

\item{COPY_VALUE}{ogs5  \strong{bc} bloc sub keyword.}

\item{DIS_TYPE}{Distribution type of boundary condition.}

\item{DIS_TYPE_CONDITION}{ogs5  \strong{bc} bloc sub keyword.}

\item{EPSILON}{ogs5  \strong{bc} bloc sub keyword.}

\item{EXCAVATION}{ogs5  \strong{bc} bloc sub keyword.}

\item{FCT_TYPE}{Name of a function defined in \strong{fct} bloc to be used as BC.}

\item{GEO_TYPE}{Name of a geometry defined in \strong{gml} to be apply the BC on.}

\item{MSH_TYPE}{ogs5  \strong{bc} bloc sub keyword.}

\item{NO_DIPS_INCREMENT}{ogs5  \strong{bc} bloc sub keyword.}

\item{PCS_TYPE}{Apply \strong{bc} for the process defined in \strong{pcs}.}

\item{PRESSURE_AS_HEAD}{ogs5  \strong{bc} bloc sub keyword.}

\item{PRIMARY_VARIABLE}{Apply \strong{bc} for the primary variable defined in \strong{pcs}.}

\item{TIME_CONTROLLED_ACTIVE}{ogs5  \strong{bc} bloc sub keyword.}

\item{TIM_TYPE}{Set name of a \strong{CURVE} defined in \strong{rfd} if bc should be
time dependend.}
}
\value{
Updated \emph{ogs5} simulation object.
}
\description{
Adds a sub-bloc to \strong{bc} bloc of \emph{ogs5} for defining boundary
conditions.
For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/bc.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))

ogs5_obj <- input_add_bc_bloc(ogs5_obj, bc_name = "BOUNDARY_CONDITION1",
                              PCS_TYPE = "GROUNDWATER_FLOW",
                              PRIMARY_VARIABLE = "HEAD",
                              GEO_TYPE = "POINT POINT0",
                              DIS_TYPE = "CONSTANT 1.0e5")
}
