% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_inputfiles.R
\name{input_add_blocs_from_file}
\alias{input_add_blocs_from_file}
\title{Add ogs5 blocs directly from existing inputfile}
\usage{
input_add_blocs_from_file(
  ogs5_obj = NULL,
  sim_basename,
  filename,
  file_dir = "",
  overwrite = FALSE
)
}
\arguments{
\item{ogs5_obj}{Instance of class \emph{ogs5}. If set to \strong{NULL}, a new ogs5
instance is created with filename and file_dir as simulation name and path
respectively.}

\item{sim_basename}{(\emph{character}) The simulation name, usually the filename of
the input files without extension.}

\item{filename}{(\emph{character}) Can be the name of the file (with extension) to be
read in, a list of file names or "all".}

\item{file_dir}{(\emph{character}) A path to the directory where the inputfiles are.}

\item{overwrite}{(\emph{logical}) If set to TRUE, existing blocs of the same type
(input file extension) will be overwritten. In order to add several blocs of
the same type, the option needs to be set to FALSE.}
}
\value{
returns an instance of \emph{ogs5} with the new input blocs added
as instances of their respective class.
}
\description{
Function to add input files directly from ogs5 input files
to an instance of class  \emph{ogs5}.
}
\details{
If \link{filename} is set to "all", the whole folder specified in
\link{file_dir} will be searched for possible input files excluding \link{*.tec},
\link{*.vtu} \link{*.vtk} and \link{*.pvd} files. However, the presence of files that are
not ogs5 input files such as for example \link{*.log}-files might have the
function crashing. For \link{filenames} that are different from \link{sim_basename}, blocs of
the class \link{ogs5_additional} are added.
}
\examples{
ex1 <- input_add_blocs_from_file(ogs5_obj = NULL,
                       sim_basename = "decal",
                       filename = "decal.mfp",
                       file_dir = "examples/benchmarks/ConcreteCrack",
                       overwrite = FALSE)
}
