% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_cct_bloc.R
\name{input_add_cct_bloc}
\alias{input_add_cct_bloc}
\title{input_add_cct_bloc}
\usage{
input_add_cct_bloc(
  x = list(),
  cct_name = character(NULL),
  MYRANK = NULL,
  NEIGHBOR = NULL,
  NNEIGHBORS = NULL
)
}
\arguments{
\item{x}{Simulation object. \emph{ogs}}

\item{cct_name}{Name of the \strong{cct} bloc.}

\item{MYRANK}{ogs5 \strong{cct} bloc sub keword.}

\item{NEIGHBOR}{ogs5 \strong{cct} bloc sub keword.}

\item{NNEIGHBORS}{ogs5 \strong{cct} bloc sub keword.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{cct} bloc of \emph{ogs5} for defining a
communication table to execute parallel runs with PETS library. For
additional documentation of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/cct.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
