% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_fct_bloc.R
\name{input_add_fct_bloc}
\alias{input_add_fct_bloc}
\title{input_add_fct_bloc}
\usage{
input_add_fct_bloc(
  x = list(),
  data_type = c("DATA", "MATRIX"),
  data_values = tibble::tibble(NULL),
  DIMENSION = NULL,
  DIS_TYPE = NULL,
  GEO_TYPE = NULL,
  TYPE = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{data_type}{'DATA' or 'MATRIX'. \emph{character}}

\item{data_values}{A \emph{tibble} with data values}

\item{DIMENSION}{Dimension if \emph{data_type} is 'Matrix'.}

\item{DIS_TYPE}{ogs5 \strong{fct} bloc sub key word.}

\item{GEO_TYPE}{Name of a geometry defined in \strong{gml} to relate the function.}

\item{TYPE}{ogs5 \strong{fct} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{fct} bloc of \emph{ogs5} for defining
functional relationships. For additional documentatiin of the input
parameters see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/fct.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} and \strong{data_values} have to be of class
\emph{character}.
}
