% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_additional.R
\name{input_add_file_as_additional_bloc}
\alias{input_add_file_as_additional_bloc}
\title{input_add_file_as_additional_bloc}
\usage{
input_add_file_as_additional_bloc(x = list(), filepath = character(NULL))
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{filepath}{Path to file to be added. \emph{character}}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a complete file as bloc to \emph{additional} sub-bloc of \emph{ogs5}
instance.
}
