% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gems_bloc.R
\name{input_add_gem_bloc}
\alias{input_add_gem_bloc}
\title{input_add_gem_bloc}
\usage{
input_add_gem_bloc(
  x = list(),
  bloc_name = "GEMS_PROPERTIES",
  CALCULATE_BOUNDARY_NODES = c("0", "1"),
  DISABLE_GEMS = FALSE,
  FLAG_COUPLING_HYDROLOGY = c("0", "1"),
  FLAG_DISABLE_GEM = NULL,
  FLAG_POROSITY_CHANGE = c("0", "1"),
  GEM_CALCULATE_BOUNDARY_NODES = NULL,
  GEM_INIT_FILE = NULL,
  GEM_THREADS = 1,
  ITERATIVE_SCHEME = FALSE,
  KINETIC_GEM = NULL,
  MAX_FAILED_NODES = NULL,
  MAX_POROSITY = NULL,
  MIN_POROSITY = NULL,
  MY_SMART_GEMS = NULL,
  PRESSURE_GEM = NULL,
  TEMPERATURE_GEM = NULL,
  TRANSPORT_B = 1
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{bloc_name}{Name of the \strong{gem} sub bloc.}

\item{CALCULATE_BOUNDARY_NODES}{\emph{character vector}}

\item{DISABLE_GEMS}{\emph{logical}}

\item{FLAG_COUPLING_HYDROLOGY}{\emph{character vector}}

\item{FLAG_DISABLE_GEM}{ogs5 \strong{gem} bloc sub key word.}

\item{FLAG_POROSITY_CHANGE}{\emph{character vector}.}

\item{GEM_CALCULATE_BOUNDARY_NODES}{ogs5 \strong{gem} bloc sub key word.}

\item{GEM_INIT_FILE}{ogs5 \strong{gem} bloc sub key word.}

\item{GEM_THREADS}{\emph{integer}}

\item{ITERATIVE_SCHEME}{\emph{logical}}

\item{KINETIC_GEM}{ogs5 \strong{gem} bloc sub key word.}

\item{MAX_FAILED_NODES}{ogs5 \strong{gem} bloc sub key word.}

\item{MAX_POROSITY}{ogs5 \strong{gem} bloc sub key word.}

\item{MIN_POROSITY}{ogs5 \strong{gem} bloc sub key word.}

\item{MY_SMART_GEMS}{ogs5 \strong{gem} bloc sub key word.}

\item{PRESSURE_GEM}{ogs5 \strong{gem} bloc sub key word.}

\item{TEMPERATURE_GEM}{ogs5 \strong{gem} bloc sub key word.}

\item{TRANSPORT_B}{\emph{logical}}
}
\value{
Updated \emph{ogs5} simulation object.
}
\description{
Adds a sub-bloc to \strong{gem} bloc of \emph{ogs5} for defining the
interface to GEMS selector code. For additional documentatoin of the input
parameters see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/gem.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
Most arguments have to be of class \emph{character}.
}
