% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gli_points.R
\name{input_add_gli_points}
\alias{input_add_gli_points}
\title{input_add_gli_points}
\usage{
input_add_gli_points(x = list(), ogs5_points = tibble::tibble())
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{ogs5_points}{\emph{tibble} with columns c("x", "y", "z", "name").}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{gli} bloc of \emph{ogs5} for defining point
geometries. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/gli/h_points}{ogs5 keyword docs}.
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))

ogs5_obj <- input_add_gli_points(x = ogs5_obj,
                ogs5_points = tibble::tibble(x = c(0, 4.7),
                                             y = c(0,0),
                                             z = c(0,0),
                                             name = c("point0", "point1")))
}
