% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_gli_points.R
\name{input_add_gli_polyline}
\alias{input_add_gli_polyline}
\title{input_add_gli_polyline}
\usage{
input_add_gli_polyline(
  x = list(),
  EPSILON = character(),
  ID = NULL,
  MAT_GROUP = NULL,
  ply_name = character(),
  POINTS = character(),
  POINT_VECTOR = NULL,
  TYPE = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{EPSILON}{Error tolerance of the specified points to the mesh nodes.}

\item{ply_name}{Name of the polyline.}

\item{POINTS}{Points to be used for the polyline.}

\item{TYPE}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{gli} bloc of \emph{ogs5} for defining a
polyline. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/gli/h_polyline}{ogs5 keyword docs}.
}
