% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_krc_bloc.R
\name{input_add_krc_bloc}
\alias{input_add_krc_bloc}
\title{input_add_krc_bloc}
\usage{
input_add_krc_bloc(
  x = list(),
  krc_name = NULL,
  mkey = NULL,
  ACTIVITY_MODEL = NULL,
  ALLOW_REACTIONS = NULL,
  BACTERIACAPACITY = NULL,
  BACTERIAL_YIELD = NULL,
  BACTERIANAME = NULL,
  BACTERIA_SPECIFIC_CAPACITY = NULL,
  BASETERM = NULL,
  CALC_SHERWOOD = NULL,
  CALC_SHERWOOD_MODIFIED = NULL,
  CHEMAPPNAME = NULL,
  COPY_CONCENTRATIONS = NULL,
  D50 = NULL,
  DEBUG_OUTPUT = FALSE,
  DM = NULL,
  DS = NULL,
  EQUATION = NULL,
  EQUILIBRIUM_CONSTANT = NULL,
  EXCHANGE_PARAMETERS = NULL,
  GAS_DISSOLUTION = NULL,
  GEOMETRY = NULL,
  GRAIN_SPHERE_RATIO = NULL,
  GROWTH = NULL,
  INHIBITIONTERMS = NULL,
  INITIAL_TIMESTEP = NULL,
  INTERFACIAL_AREA = NULL,
  ISOTOPE_FRACTIONATION = NULL,
  LAGNEAU_BENCHMARK = FALSE,
  LENGTH = NULL,
  MECHANISMTERM = NULL,
  MICROBENAME = NULL,
  MINERALNAME = NULL,
  MIN_BACTERIACONC = NULL,
  MIN_CONCENTRATION_REPLACE = NULL,
  MIN_TIMESTEP = NULL,
  MONODTERMS = NULL,
  MONOD_REACTION_NAME = NULL,
  NAME = NULL,
  NAPL_CONTENT_INI = NULL,
  NAPL_CONTENT_RES = NULL,
  NAPL_PROPERTIES = NULL,
  NO_REACTIONS = NULL,
  OMEGA_THRESHOLD = NULL,
  PRECIPITATION_BY_BASETERM_ONLY = FALSE,
  PRECIPITATION_EXPONENT = NULL,
  PRECIPITATION_FACTOR = NULL,
  PRODUCTIONSTOCH = NULL,
  PRODUCTIONTERMS = NULL,
  RATECONSTANT = NULL,
  RATE_EXPONENTS = NULL,
  REACTION_DEACTIVATION = NULL,
  REACTION_ORDER = NULL,
  REACTIVE_SURFACE_AREA = NULL,
  RELATIVE_ERROR = NULL,
  SCALE_DCDT = FALSE,
  SHERWOOD_MODEL = NULL,
  SOLVER_TYPE = NULL,
  SORPTION_TYPE = NULL,
  SORT_NODES = FALSE,
  STANDARD_GIBBS_ENERGY = NULL,
  SURFACES = NULL,
  SWITCH_OFF_GEOMETRY = NULL,
  TEMPERATURE_DEPENDENCE = NULL,
  THRESHHOLDTERMS = NULL,
  TORTUOSITY = NULL,
  TYPE = NULL,
  UI = NULL,
  drmc = NULL,
  drmc_PARAMETERS = NULL
)
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{krc_name}{Name of the \strong{krc} sub bloc.}

\item{mkey}{Main keyword of the \emph{ogs5_krc_bloc}: c("BLOB_PROPERTIES",
"KINREACTIONDATA", "MICROBE_PROPERTIES", "REACTION").}

\item{ACTIVITY_MODEL}{ogs5 \strong{krc} bloc sub key word.}

\item{ALLOW_REACTIONS}{ogs5 \strong{krc} bloc sub key word.}

\item{BACTERIACAPACITY}{ogs5 \strong{krc} bloc sub key word.}

\item{BACTERIAL_YIELD}{ogs5 \strong{krc} bloc sub key word.}

\item{BACTERIANAME}{ogs5 \strong{krc} bloc sub key word.}

\item{BACTERIA_SPECIFIC_CAPACITY}{ogs5 \strong{krc} bloc sub key word.}

\item{BASETERM}{ogs5 \strong{krc} bloc sub key word.}

\item{CALC_SHERWOOD}{ogs5 \strong{krc} bloc sub key word.}

\item{CALC_SHERWOOD_MODIFIED}{ogs5 \strong{krc} bloc sub key word.}

\item{CHEMAPPNAME}{ogs5 \strong{krc} bloc sub key word.}

\item{COPY_CONCENTRATIONS}{ogs5 \strong{krc} bloc sub key word.}

\item{D50}{ogs5 \strong{krc} bloc sub key word.}

\item{DEBUG_OUTPUT}{ogs5 \strong{krc} bloc sub key word.}

\item{DM}{ogs5 \strong{krc} bloc sub key word.}

\item{DS}{ogs5 \strong{krc} bloc sub key word.}

\item{EQUATION}{ogs5 \strong{krc} bloc sub key word.}

\item{EQUILIBRIUM_CONSTANT}{ogs5 \strong{krc} bloc sub key word.}

\item{EXCHANGE_PARAMETERS}{ogs5 \strong{krc} bloc sub key word.}

\item{GAS_DISSOLUTION}{ogs5 \strong{krc} bloc sub key word.}

\item{GEOMETRY}{ogs5 \strong{krc} bloc sub key word.}

\item{GRAIN_SPHERE_RATIO}{ogs5 \strong{krc} bloc sub key word.}

\item{GROWTH}{ogs5 \strong{krc} bloc sub key word.}

\item{INHIBITIONTERMS}{ogs5 \strong{krc} bloc sub key word.}

\item{INITIAL_TIMESTEP}{ogs5 \strong{krc} bloc sub key word.}

\item{INTERFACIAL_AREA}{ogs5 \strong{krc} bloc sub key word.}

\item{ISOTOPE_FRACTIONATION}{ogs5 \strong{krc} bloc sub key word.}

\item{LAGNEAU_BENCHMARK}{ogs5 \strong{krc} bloc sub key word.}

\item{LENGTH}{ogs5 \strong{krc} bloc sub key word.}

\item{MECHANISMTERM}{ogs5 \strong{krc} bloc sub key word.}

\item{MICROBENAME}{ogs5 \strong{krc} bloc sub key word.}

\item{MINERALNAME}{ogs5 \strong{krc} bloc sub key word.}

\item{MIN_BACTERIACONC}{ogs5 \strong{krc} bloc sub key word.}

\item{MIN_CONCENTRATION_REPLACE}{ogs5 \strong{krc} bloc sub key word.}

\item{MIN_TIMESTEP}{ogs5 \strong{krc} bloc sub key word.}

\item{MONODTERMS}{ogs5 \strong{krc} bloc sub key word.}

\item{MONOD_REACTION_NAME}{ogs5 \strong{krc} bloc sub key word.}

\item{NAME}{ogs5 \strong{krc} bloc sub key word.}

\item{NAPL_CONTENT_INI}{ogs5 \strong{krc} bloc sub key word.}

\item{NAPL_CONTENT_RES}{ogs5 \strong{krc} bloc sub key word.}

\item{NAPL_PROPERTIES}{ogs5 \strong{krc} bloc sub key word.}

\item{NO_REACTIONS}{ogs5 \strong{krc} bloc sub key word.}

\item{OMEGA_THRESHOLD}{ogs5 \strong{krc} bloc sub key word.}

\item{PRECIPITATION_BY_BASETERM_ONLY}{ogs5 \strong{krc} bloc sub key word.}

\item{PRECIPITATION_EXPONENT}{ogs5 \strong{krc} bloc sub key word.}

\item{PRECIPITATION_FACTOR}{ogs5 \strong{krc} bloc sub key word.}

\item{PRODUCTIONSTOCH}{ogs5 \strong{krc} bloc sub key word.}

\item{PRODUCTIONTERMS}{ogs5 \strong{krc} bloc sub key word.}

\item{RATECONSTANT}{ogs5 \strong{krc} bloc sub key word.}

\item{RATE_EXPONENTS}{ogs5 \strong{krc} bloc sub key word.}

\item{REACTION_DEACTIVATION}{ogs5 \strong{krc} bloc sub key word.}

\item{REACTION_ORDER}{ogs5 \strong{krc} bloc sub key word.}

\item{REACTIVE_SURFACE_AREA}{ogs5 \strong{krc} bloc sub key word.}

\item{RELATIVE_ERROR}{ogs5 \strong{krc} bloc sub key word.}

\item{SCALE_DCDT}{ogs5 \strong{krc} bloc sub key word.}

\item{SHERWOOD_MODEL}{ogs5 \strong{krc} bloc sub key word.}

\item{SOLVER_TYPE}{ogs5 \strong{krc} bloc sub key word.}

\item{SORPTION_TYPE}{ogs5 \strong{krc} bloc sub key word.}

\item{SORT_NODES}{ogs5 \strong{krc} bloc sub key word.}

\item{STANDARD_GIBBS_ENERGY}{ogs5 \strong{krc} bloc sub key word.}

\item{SURFACES}{ogs5 \strong{krc} bloc sub key word.}

\item{SWITCH_OFF_GEOMETRY}{ogs5 \strong{krc} bloc sub key word.}

\item{TEMPERATURE_DEPENDENCE}{ogs5 \strong{krc} bloc sub key word.}

\item{THRESHHOLDTERMS}{ogs5 \strong{krc} bloc sub key word.}

\item{TORTUOSITY}{ogs5 \strong{krc} bloc sub key word.}

\item{TYPE}{ogs5 \strong{krc} bloc sub key word.}

\item{UI}{ogs5 \strong{krc} bloc sub key word.}

\item{drmc}{ogs5 \strong{krc} bloc sub key word.}

\item{drmc_PARAMETERS}{ogs5 \strong{krc} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{krc} bloc of \emph{ogs5} for defining kinetic
reactions. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/krc.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
Most arguments have to be of class \emph{character}.
}
