% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_mcp_bloc.R
\name{input_add_mcp_bloc}
\alias{input_add_mcp_bloc}
\title{input_add_mcp_bloc}
\usage{
input_add_mcp_bloc(
  x = list(),
  ACENTRIC_FACTOR = NULL,
  A_ZERO = NULL,
  BUBBLE_VELOCITY = NULL,
  CRITICAL_PRESSURE = NULL,
  CRITICAL_TEMPERATURE = NULL,
  DECAY = NULL,
  DIFFUSION = NULL,
  FLUID_ID = NULL,
  FLUID_PHASE = NULL,
  FORMULA = NULL,
  ISOTHERM = NULL,
  MAXIMUM_AQUEOUS_SOLUBILITY = NULL,
  MINERAL_DENSITY = NULL,
  MOBILE = NULL,
  MOLAR_DENSITY = NULL,
  MOLAR_VOLUME = NULL,
  MOLAR_WEIGHT = NULL,
  MOL_MASS = NULL,
  NAME = NULL,
  OutputMassOfComponentInModel = NULL,
  TRANSPORT_PHASE = NULL,
  VALENCE = NULL,
  VOLUME_DIFFUSION = NULL
)
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{ACENTRIC_FACTOR}{ogs5 \strong{mcp} bloc sub key word.}

\item{A_ZERO}{ogs5 \strong{mcp} bloc sub key word.}

\item{BUBBLE_VELOCITY}{ogs5 \strong{mcp} bloc sub key word.}

\item{CRITICAL_PRESSURE}{ogs5 \strong{mcp} bloc sub key word.}

\item{CRITICAL_TEMPERATURE}{ogs5 \strong{mcp} bloc sub key word.}

\item{DECAY}{ogs5 \strong{mcp} bloc sub key word.}

\item{DIFFUSION}{ogs5 \strong{mcp} bloc sub key word.}

\item{FLUID_ID}{ogs5 \strong{mcp} bloc sub key word.}

\item{FLUID_PHASE}{ogs5 \strong{mcp} bloc sub key word.}

\item{FORMULA}{ogs5 \strong{mcp} bloc sub key word.}

\item{ISOTHERM}{ogs5 \strong{mcp} bloc sub key word.}

\item{MAXIMUM_AQUEOUS_SOLUBILITY}{ogs5 \strong{mcp} bloc sub key word.}

\item{MINERAL_DENSITY}{ogs5 \strong{mcp} bloc sub key word.}

\item{MOBILE}{ogs5 \strong{mcp} bloc sub key word.}

\item{MOLAR_DENSITY}{ogs5 \strong{mcp} bloc sub key word.}

\item{MOLAR_VOLUME}{ogs5 \strong{mcp} bloc sub key word.}

\item{MOLAR_WEIGHT}{ogs5 \strong{mcp} bloc sub key word.}

\item{MOL_MASS}{ogs5 \strong{mcp} bloc sub key word.}

\item{NAME}{Component name.}

\item{OutputMassOfComponentInModel}{ogs5 \strong{mcp} bloc sub key word.}

\item{TRANSPORT_PHASE}{ogs5 \strong{mcp} bloc sub key word.}

\item{VALENCE}{ogs5 \strong{mcp} bloc sub key word.}

\item{VOLUME_DIFFUSION}{ogs5 \strong{mcp} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{mcp} bloc of \emph{ogs5} for defining component
properties. For additional documentatoin of the input parameters
see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/mcp.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
ogs5_obj <- input_add_mcp_bloc(ogs5_obj, NAME = "C(4)",
                               MOBILE = "1",
                               DIFFUSION = "1 0.0e-9",
                               VALENCE = "-2")
}
