% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_mfp_bloc.R
\name{input_add_mfp_bloc}
\alias{input_add_mfp_bloc}
\title{input_add_mfp_bloc}
\usage{
input_add_mfp_bloc(
  x = list(),
  COMPONENTS = NULL,
  COMPRESSIBILITY = NULL,
  DAT_TYPE = NULL,
  DECAY = NULL,
  DENSITY = character(NULL),
  DIFFUSION = NULL,
  DRHO_DT_UNSATURATED = NULL,
  EOS_TYPE = NULL,
  FLUID_NAME = character(NULL),
  FLUID_TYPE = character(NULL),
  GRAVITY = NULL,
  HEAT_CONDUCTIVITY = NULL,
  ISOTHERM = NULL,
  JTC = NULL,
  NON_GRAVITY = NULL,
  PHASE_DIFFUSION = NULL,
  SPECIFIC_HEAT_CAPACITY = NULL,
  SPECIFIC_HEAT_SOURCE = NULL,
  TEMPERATURE = NULL,
  VISCOSITY = character(NULL)
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{VISCOSITY}{}
}
\value{

}
\description{
Adds a sub-bloc to \strong{mfp} bloc of \emph{ogs5} for defining fluid
properties. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/mfp.html}{ogs5 keyword docs}
}
\examples{
ogs5_obj <- input_add_mfp_bloc(ogs5_obj, FLUID_NAME = "FLUID_PROPERTIES1",
                               FLUID_TYPE = "LIQUID",
                               DENSITY = "1 1000.0",
                               VISCOSITY = "1 1.0e-3",
                               HEAT_CONDUCTIVITY = "1 0.0")
}
