% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_msh.R
\name{input_add_msh_bloc}
\alias{input_add_msh_bloc}
\title{input_add_msh_bloc}
\usage{
input_add_msh_bloc(
  x = list(),
  msh_name = character(),
  AREA = NULL,
  AXISYMMETRY = NULL,
  CROSS_SECTION = NULL,
  ELEMENTS = tibble::tibble(),
  GEO_NAME = NULL,
  GEO_TYPE = NULL,
  LAYER = NULL,
  NODES = tibble::tibble(),
  PCS_TYPE = "NO_PCS"
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{msh_name}{Name of the mesh. \emph{character}}

\item{AREA}{ogs5 \strong{msh} bloc sub key word.}

\item{AXISYMMETRY}{ogs5 \strong{msh} bloc sub key word.}

\item{CROSS_SECTION}{ogs5 \strong{msh} bloc sub key word.}

\item{ELEMENTS}{A \emph{tibble} with columns material_id ele_type node1 node2
(node3, node4)}

\item{GEO_NAME}{ogs5 \strong{msh} bloc sub key word.}

\item{GEO_TYPE}{ogs5 \strong{msh} bloc sub key word.}

\item{LAYER}{ogs5 \strong{msh} bloc sub key word.}

\item{NODES}{A \emph{tibble} with columns x, y, z.}

\item{PCS_TYPE}{ogs5 \strong{msh} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{msh} bloc of \emph{ogs5} for defining meshes.
Only simple geometric meshes such as lines,
rectangles and cubes can be created here. For more complicated geometries
please refer to third party software such as \strong{gmsh}.
For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/msh.html}{ogs5 keyword docs}.
}
\examples{
mesh_lst <- create_structured_mesh_nodes_ele(lx = 4.7, nx = 94)
ogs5_obj <- input_add_msh_bloc(x = ogs5_obj, msh_name = "base_mesh",
                               NODES = mesh_lst[[1]],
                               ELEMENTS = mesh_lst[[2]])
}
