% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_num_bloc.R
\name{input_add_num_bloc}
\alias{input_add_num_bloc}
\title{input_add_num_bloc}
\usage{
input_add_num_bloc(
  x = list(),
  num_name = NULL,
  COUPLED_PROCESS = NULL,
  COUPLING_CONTROL = NULL,
  COUPLING_ITERATIONS = NULL,
  DYNAMIC_DAMPING = NULL,
  ELE_GAUSS_POINTS = NULL,
  ELE_MASS_LUMPING = NULL,
  ELE_SUPG = NULL,
  ELE_UPWINDING = NULL,
  EXTERNAL_SOLVER_OPTION = NULL,
  FEM_FCT = NULL,
  GRAVITY_PROFILE = NULL,
  LINEAR_SOLVER = NULL,
  LOCAL_PICARD1 = NULL,
  NON_LINEAR_ITERATION = NULL,
  NON_LINEAR_SOLVER = NULL,
  NON_LINEAR_UPDATE_VELOCITY = NULL,
  OVERALL_COUPLING = NULL,
  PCS_TYPE = NULL,
  PLASTICITY_TOLERANCE = NULL,
  RENUMBER = NULL,
  TIME_STEPS = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{num_name}{Name of the sub-bloc.}

\item{COUPLED_PROCESS}{ogs5 \strong{num} bloc sub key word.}

\item{COUPLING_CONTROL}{ogs5 \strong{num} bloc sub key word.}

\item{COUPLING_ITERATIONS}{ogs5 \strong{num} bloc sub key word.}

\item{DYNAMIC_DAMPING}{ogs5 \strong{num} bloc sub key word.}

\item{ELE_GAUSS_POINTS}{ogs5 \strong{num} bloc sub key word.}

\item{ELE_MASS_LUMPING}{ogs5 \strong{num} bloc sub key word.}

\item{ELE_SUPG}{ogs5 \strong{num} bloc sub key word.}

\item{ELE_UPWINDING}{ogs5 \strong{num} bloc sub key word.}

\item{EXTERNAL_SOLVER_OPTION}{ogs5 \strong{num} bloc sub key word.}

\item{FEM_FCT}{ogs5 \strong{num} bloc sub key word.}

\item{GRAVITY_PROFILE}{ogs5 \strong{num} bloc sub key word.}

\item{LINEAR_SOLVER}{ogs5 \strong{num} bloc sub key word.}

\item{LOCAL_PICARD1}{ogs5 \strong{num} bloc sub key word.}

\item{NON_LINEAR_ITERATION}{ogs5 \strong{num} bloc sub key word.}

\item{NON_LINEAR_SOLVER}{ogs5 \strong{num} bloc sub key word.}

\item{NON_LINEAR_UPDATE_VELOCITY}{ogs5 \strong{num} bloc sub key word.}

\item{OVERALL_COUPLING}{ogs5 \strong{num} bloc sub key word.}

\item{PCS_TYPE}{ogs5 \strong{num} bloc sub key word.}

\item{PLASTICITY_TOLERANCE}{ogs5 \strong{num} bloc sub key word.}

\item{RENUMBER}{ogs5 \strong{num} bloc sub key word.}

\item{TIME_STEPS}{ogs5 \strong{num} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{num} bloc of \emph{ogs5} for defining numerical
solver properties. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/num.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))

ogs5_obj <- input_add_num_bloc(ogs5_obj, num_name = "NUMERICS1",
                               PCS_TYPE = "GROUNDWATER_FLOW",
                               ELE_GAUSS_POINTS = "3",
                               LINEAR_SOLVER = "2 6 1.e-014 1000 1.0 1 2")
}
