% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_out.R
\name{input_add_out_bloc}
\alias{input_add_out_bloc}
\title{input_add_out_bloc}
\usage{
input_add_out_bloc(
  x = list(),
  out_name = NULL,
  AMPLIFIER = NULL,
  DAT_TYPE = NULL,
  DIS_TYPE = NULL,
  ELE_VALUES = NULL,
  GEO_TYPE = NULL,
  MFP_VALUES = NULL,
  MMP_VALUES = NULL,
  MSH_TYPE = NULL,
  NOD_VALUES = NULL,
  PCON_VALUES = NULL,
  PCS_TYPE = NULL,
  RWPT_VALUES = NULL,
  TECPLOT_ZONE_SHARE = FALSE,
  TIM_TYPE = NULL,
  VARIABLESHARING = FALSE
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{out_name}{Name of the \strong{out} sub-bloc. \emph{character}}

\item{AMPLIFIER}{ogs5 \strong{out} bloc sub key word.}

\item{DAT_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{DIS_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{ELE_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{GEO_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{MFP_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{MMP_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{MSH_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{NOD_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{PCON_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{PCS_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{RWPT_VALUES}{ogs5 \strong{out} bloc sub key word.}

\item{TECPLOT_ZONE_SHARE}{\emph{logical}}

\item{TIM_TYPE}{ogs5 \strong{out} bloc sub key word.}

\item{VARIABLESHARING}{\emph{logical}}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{out} bloc of \emph{ogs5} for defining
simulation output conditions. For additional documentatoin of the input
parameters see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/out.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
Most arguments have to be of class \emph{character}.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))

ogs5_obj <- input_add_out_bloc(ogs5_obj, out_name = "OUTPUT1",
                   NOD_VALUES = "C(4) Ca Mg Cl pH pe Calcite Dolomite(dis)",
                   GEO_TYPE = "POINT POINT2",
                   DAT_TYPE = "TECPLOT")
}
