% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_pcs_bloc.R
\name{input_add_pcs_bloc}
\alias{input_add_pcs_bloc}
\title{input_add_pcs_bloc}
\usage{
input_add_pcs_bloc(
  x = list(),
  pcs_name = NULL,
  APP_TYPE = NULL,
  BOUNDARY_CONDITION_OUTPUT = FALSE,
  COUNT = NULL,
  CPL_TYPE = NULL,
  DEACTIVATED_SUBDOMAIN = NULL,
  DISSOLVED_CO2_INGAS_PCS_NAME = NULL,
  DISSOLVED_CO2_PCS_NAME = NULL,
  ELEMENT_MATRIX_OUTPUT = NULL,
  GEO_TYPE = NULL,
  MEDIUM_TYPE = NULL,
  MEMORY_TYPE = NULL,
  MSH_TYPE = NULL,
  NEGLECT_H_INI_EFFECT = NULL,
  NUM_TYPE = NULL,
  OutputMassOfGasInModel = FALSE,
  PCS_TYPE = NULL,
  PHASE_TRANSITION = NULL,
  PRIMARY_VARIABLE = NULL,
  PROCESSED_BC = NULL,
  RELOAD = NULL,
  SATURATION_SWITCH = FALSE,
  SAVE_ECLIPSE_DATA_FILES = FALSE,
  SIMULATOR = NULL,
  SIMULATOR_MODEL_PATH = NULL,
  SIMULATOR_PATH = NULL,
  SIMULATOR_WELL_PATH = NULL,
  ST_RHS = NULL,
  TIME_CONTROLLED_EXCAVATION = NULL,
  TIM_TYPE = NULL,
  UPDATE_INI_STATE = NULL,
  USE_PRECALCULATED_FILES = FALSE,
  USE_VELOCITIES_FOR_TRANSPORT = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{pcs_name}{Name of the \strong{pcs} sub-bloc.}

\item{APP_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{BOUNDARY_CONDITION_OUTPUT}{ogs5 \strong{pcs} bloc sub key word.}

\item{COUNT}{ogs5 \strong{pcs} bloc sub key word.}

\item{CPL_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{DEACTIVATED_SUBDOMAIN}{ogs5 \strong{pcs} bloc sub key word.}

\item{DISSOLVED_CO2_INGAS_PCS_NAME}{ogs5 \strong{pcs} bloc sub key word.}

\item{DISSOLVED_CO2_PCS_NAME}{ogs5 \strong{pcs} bloc sub key word.}

\item{ELEMENT_MATRIX_OUTPUT}{ogs5 \strong{pcs} bloc sub key word.}

\item{GEO_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{MEDIUM_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{MEMORY_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{MSH_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{NEGLECT_H_INI_EFFECT}{ogs5 \strong{pcs} bloc sub key word.}

\item{NUM_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{OutputMassOfGasInModel}{ogs5 \strong{pcs} bloc sub key word.}

\item{PCS_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{PHASE_TRANSITION}{ogs5 \strong{pcs} bloc sub key word.}

\item{PRIMARY_VARIABLE}{ogs5 \strong{pcs} bloc sub key word.}

\item{PROCESSED_BC}{ogs5 \strong{pcs} bloc sub key word.}

\item{RELOAD}{ogs5 \strong{pcs} bloc sub key word.}

\item{SATURATION_SWITCH}{ogs5 \strong{pcs} bloc sub key word.}

\item{SAVE_ECLIPSE_DATA_FILES}{ogs5 \strong{pcs} bloc sub key word.}

\item{SIMULATOR}{ogs5 \strong{pcs} bloc sub key word.}

\item{SIMULATOR_MODEL_PATH}{ogs5 \strong{pcs} bloc sub key word.}

\item{SIMULATOR_PATH}{ogs5 \strong{pcs} bloc sub key word.}

\item{SIMULATOR_WELL_PATH}{ogs5 \strong{pcs} bloc sub key word.}

\item{ST_RHS}{ogs5 \strong{pcs} bloc sub key word.}

\item{TIME_CONTROLLED_EXCAVATION}{ogs5 \strong{pcs} bloc sub key word.}

\item{TIM_TYPE}{ogs5 \strong{pcs} bloc sub key word.}

\item{UPDATE_INI_STATE}{ogs5 \strong{pcs} bloc sub key word.}

\item{USE_PRECALCULATED_FILES}{ogs5 \strong{pcs} bloc sub key word.}

\item{USE_VELOCITIES_FOR_TRANSPORT}{ogs5 \strong{pcs} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{pcs} bloc of \emph{ogs5} for defining processes
For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/pcs.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
Most arguments have to be of class \emph{character}.
}
\examples{
ogs5_obj <- input_add_pcs_bloc(ogs5_obj, pcs_name = "PROCESS1",
                               PCS_TYPE = "GROUNDWATER_FLOW",
                               NUM_TYPE = "NEW",
                               ELEMENT_MATRIX_OUTPUT = "0")
}
