% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_pqc.R
\name{input_add_pqc_skeybloc}
\alias{input_add_pqc_skeybloc}
\title{input_add_pqc_skeybloc}
\usage{
input_add_pqc_skeybloc(
  x = list(),
  skey = c("EQUILIBRIUM_PHASES", "EXCHANGE", "GAS_PHASE", "ISOTOPES", "KINETICS",
    "KNOBS", "RATES", "REACTION", "SOLID_SOLUTION", "SOLUTION", "SURFACE"),
  bloc_text = character(NULL)
)
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{skey}{Phreeqc keyword: c("EQUILIBRIUM_PHASES", "EXCHANGE", "GAS_PHASE",
"ISOTOPES", "KINETICS", "KNOBS", "RATES", "REACTION", "SOLID_SOLUTION",
"SOLUTION", "SURFACE"). For Phreeqc documentation \link{see}
(https://water.usgs.gov/water-resources/software/PHREEQC/documentation/phreeqc3-html/phreeqc3.htm)}

\item{bloc_text}{\emph{character} defining the sub-bloc.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a keyword specific \strong{pqc} sub-bloc to an \emph{ogs5} simulation
object to define the input of the chemical solver \link{Phreeqc}
(https://www.usgs.gov/software/phreeqc-version-3/). For further
documentation see \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/pqc.html}{ogs5 keyword docs}
}
