% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_pqc.R
\name{input_add_pqcfile_as_pqc_bloc}
\alias{input_add_pqcfile_as_pqc_bloc}
\title{input_add_pqcfile_as_pqc_bloc}
\usage{
input_add_pqcfile_as_pqc_bloc(x = list(), pqc_filename = character(NULL))
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{pqc_filename}{\emph{character} Path to the *\strong{.pqc} input file.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Add an entire *\strong{.pqc} input file as \strong{pqc} bloc to an \emph{ogs5}
simulation object to define the input of the chemical solver \href{https://www.usgs.gov/software/phreeqc-version-3/}{Phreeqc}. For further
documentation see \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/pqc.html}{ogs5 keyword docs}
}
