% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_rfd_bloc.R
\name{input_add_rfd_bloc}
\alias{input_add_rfd_bloc}
\title{input_add_rfd_bloc}
\usage{
input_add_rfd_bloc(
  x = list(),
  rfd_name = NULL,
  mkey = NULL,
  data = tibble::tibble(NULL),
  INTERPOLATION = NULL,
  MSH_TYPE = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{rfd_name}{\emph{character} Name of the \strong{rfd} sub-bloc.}

\item{mkey}{\emph{character} ogs5 #-keyword: c("CURVE", "CURVES").}

\item{data}{\emph{Tibble} with two \emph{numeric} columns. One colum has to be named
"time".}

\item{INTERPOLATION}{ogs5 \strong{rfd} bloc sub key word.}

\item{MSH_TYPE}{ogs5 \strong{rfd} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{rfd} bloc of \emph{ogs5} for defining time
dependent functions. For additional documentation of the input parameters
see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/rfd.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
}
