% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_tim_bloc.R
\name{input_add_tim_bloc}
\alias{input_add_tim_bloc}
\title{input_add_tim_bloc}
\usage{
input_add_tim_bloc(
  x = list(),
  tim_name = character(NULL),
  CRITICAL_TIME = NULL,
  INDEPENDENT = NULL,
  PCS_TYPE = character(NULL),
  SUBSTEPS = NULL,
  TIME_CONTROL = NULL,
  TIME_END = numeric(NULL),
  TIME_FIXED_POINTS = NULL,
  TIME_SPLITS = NULL,
  TIME_START = numeric(NULL),
  TIME_STEPS = NULL,
  TIME_UNIT = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{tim_name}{Name of the \strong{tim} sub-bloc.}

\item{TIME_UNIT}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{tim} bloc of \emph{ogs5} for defining numerical
time discritization. For additional documentatoin of the input parameters
see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/tim.html}{ogs5 keyword docs}.
}
\examples{
ogs5_obj <- input_add_tim_bloc(ogs5_obj, tim_name = "TIME_STEPPING1",
                              PCS_TYPE = "GROUNDWATER_FLOW",
                              TIME_STEPS = "210 100.0",
                              TIME_END = "21000.0",
                              TIME_START = "0.0")
}
