% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_vtu.R
\name{ogs5_extract_time_from_vtu}
\alias{ogs5_extract_time_from_vtu}
\title{ogs5_extract_time_from_vtu}
\usage{
ogs5_extract_time_from_vtu(filename)
}
\arguments{
\item{filename}{Path to file. \emph{character}}
}
\value{
\emph{numeric} vector.
}
\description{
Extract timesteps and time from *\strong{.vtu} file.
}
