% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_keywords.R
\name{ogs5_get_keywordlist}
\alias{ogs5_get_keywordlist}
\title{ogs5_get_keywordlist}
\usage{
ogs5_get_keywordlist()
}
\value{
\emph{List} of keywords.
}
\description{
Get all \strong{OpenGeoSys-v5} keywords as \emph{list}.
Function takes no arguments.
}
