% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_get_output.R
\name{ogs5_get_output_specific}
\alias{ogs5_get_output_specific}
\title{ogs5_get_output_specific}
\usage{
ogs5_get_output_specific(
  ogs5 = list(),
  outbloc_names = character(),
  out_filepath = NULL
)
}
\arguments{
\item{ogs5}{\emph{ogs5} simulation object.}

\item{outbloc_names}{\emph{Character} vector of the \strong{out} sub bloc names.}

\item{out_filepath}{Path to output files. \emph{character}. Default:
'attributes(ogs5)$sim_path'.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Import the output from an \strong{ogs5} simulation defined in a
specific \emph{ogs5} \strong{out} sub-bloc. Add output to \emph{ogs5} object.
}
\examples{
ex1 <- ogs5_get_output_specific(ex1, outbloc_names = "tracer")
}
