% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_list_output}
\alias{ogs5_list_output}
\title{ogs5_list_output}
\usage{
ogs5_list_output(ogs5_sublist, ...)
}
\arguments{
\item{ogs5_sublist}{\emph{ogs5} sub bloc.}

\item{...}{Additional arguments to pass.}
}
\description{
Output class specific sub bloc of \emph{ogs5} object.
}
