% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_print_fct_bloc}
\alias{ogs5_print_fct_bloc}
\title{ogs5_print_fct_bloc}
\usage{
ogs5_print_fct_bloc(mkey_bloc, mkey)
}
\arguments{
\item{mkey_bloc}{\emph{ogs5} main input bloc.}

\item{mkey}{Respective \strong{ogs5} main keyword. \emph{character}}
}
\description{
Output \emph{ogs5} \strong{FCT} bloc.
}
