% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_print_mkey_bloc}
\alias{ogs5_print_mkey_bloc}
\title{ogs5_print_mkey_bloc}
\usage{
ogs5_print_mkey_bloc(mkey_bloc = list(), mkey = character(NULL), stack = FALSE)
}
\arguments{
\item{mkey_bloc}{\emph{ogs5} main input bloc.}

\item{mkey}{Respective \strong{ogs5} main keyword. \emph{character}}

\item{stack}{If output should be on seperate lines. \emph{logical}}
}
\description{
Output \emph{ogs5} main keyword bloc.
}
