% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_print_msh_mkey_bloc}
\alias{ogs5_print_msh_mkey_bloc}
\title{ogs5_print_msh_mkey_bloc}
\usage{
ogs5_print_msh_mkey_bloc(mkey_bloc = list())
}
\arguments{
\item{mkey_bloc}{\emph{ogs5} \strong{MSH} input bloc.}
}
\description{
Output \emph{ogs5} \strong{MSH} bloc.
}
