% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_process_imported_vtu_data.R
\name{ogs5_read_data_at_nodes}
\alias{ogs5_read_data_at_nodes}
\title{ogs5_read_data_at_nodes}
\usage{
ogs5_read_data_at_nodes(
  ogs5 = list(),
  outbloc_name = character(),
  node_coords = NULL
)
}
\arguments{
\item{ogs5}{\emph{ogs5} simulation object.}

\item{outbloc_name}{Name of the corresponding \emph{ogs5} \strong{out} sub-bloc where
the output was defined.}

\item{node_coords}{\emph{Tibble} of coordinates (x,y,z) of nodes where data has
to be extracted.}
}
\value{
\emph{Tibble}
}
\description{
Extract data from \emph{lists} read from *\strong{.vtu} files
at given node locations.
}
\examples{
foo <- ogs5_read_data_at_nodes(ex1, outbloc_name = "waterflow",
           node_coords = tibble::tibble(x=c(2,5), y=c(0,0), z=c(0,0)))
}
