% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_many_tecplots}
\alias{ogs5_read_many_tecplots}
\title{ogs5_read_many_tecplots}
\usage{
ogs5_read_many_tecplots(filepath = character(), geo_object = character())
}
\arguments{
\item{filepath}{(\emph{character}) path to \link{*.tec}-files.}

\item{geo_object}{(\emph{character}) Name for the type of geometric ogs object
the tec file is associated with. Can be any of "POINT", "POLYLINE", "
SURFACE", "domain".}
}
\value{
\emph{Data.frame} with timeseries of the simulation reasults.
}
\description{
Function to read in any \link{*.tec}-files inside a directory. Calls on the
functions \code{ogs5_read_tecplot_point}, \code{ogs5_read_tecplot_polyline},
\code{ogs5_read_tecplot_domain} depending on the specified \emph{geo_object} type.
}
