% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_tecplot_domain}
\alias{ogs5_read_tecplot_domain}
\title{ogs5_read_tecplot_domain}
\usage{
ogs5_read_tecplot_domain(filepath)
}
\arguments{
\item{filepath}{path to filename}
}
\value{

}
\description{
ogs5_read_tecplot_domain
}
