% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_tecplot_point}
\alias{ogs5_read_tecplot_point}
\title{ogs5_read_tecplot_point}
\usage{
ogs5_read_tecplot_point(filename = character())
}
\arguments{
\item{filename}{Path to *\strong{.tec} file. \emph{character}}
}
\value{
Dataframe.
}
\description{
Read data from *\strong{.tec} file of a point.
}
