% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_run_sim.R
\name{ogs5_run}
\alias{ogs5_run}
\title{ogs5_run}
\usage{
ogs5_run(
  ogs5_obj = list(),
  ogs_exe = NULL,
  run_path = NULL,
  log_output = TRUE,
  log_path = NULL,
  use_mpi = FALSE,
  number_of_cores = NULL,
  wait = TRUE
)
}
\arguments{
\item{ogs5_obj}{\emph{ogs5} simulation object.}

\item{ogs_exe}{Path to \strong{ogs5} executable.}

\item{run_path}{Path to directory where the simulation should be executed.
Default: use 'attributes(ogs5_obj)$sim_path'}

\item{log_output}{Should a log file be written? \emph{logical}}

\item{log_path}{Path of directory to write log file. \emph{character}}

\item{use_mpi}{Execute parallelized simulation using MPI. \emph{logical}}

\item{number_of_cores}{Use number of computing cores if 'use_mpi=TRUE'.}

\item{wait}{if 'TRUE', Freeze R-session until simulation finished. \emph{logical}}
}
\description{
Run \strong{ogs5} simulation via call to OS.
}
\examples{
 ogs5_run(ogs5_obj = sim1, ogs_exe = "inst/ogs/ogs_5.76",
          run_path = NULL,
          log_output = TRUE,
          log_path = "examples/tmp/ex1/log")
}
