% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_write_inputfiles}
\alias{ogs5_write_inputfiles}
\title{ogs5_write_inputfiles}
\usage{
ogs5_write_inputfiles(ogs5_obj = list(), type = "all", folderpath = NULL)
}
\arguments{
\item{ogs5_obj}{\emph{ogs5} simulation object}

\item{type}{\emph{character} Type of the \emph{ogs5} main bloc to be written. c("all", "bc", "cct",
"fct", "gem", "gli", "ic", "krc", "mcp", "mfp", "mmp", "msh", "msp", num",
"out", "pcs", "pqc", "rei", "st", "tim", "rfd", "additional")}

\item{folderpath}{\emph{character} Path to write the outpufile(s) to.}
}
\description{
Export \emph{ogs5} object and write ogs5 input files.
}
\examples{
ogs5_write_inputfiles(ogs5_obj, "all")
}
