% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_write_ogs5_additional}
\alias{ogs5_write_ogs5_additional}
\title{ogs5_write_ogs5_additional}
\usage{
ogs5_write_ogs5_additional(folderpath, ogs5_list)
}
\arguments{
\item{folderpath}{Path to directory where to write out the file.}

\item{ogs5_list}{\emph{ogs5} \strong{additional} bloc.}
}
\description{
Output \emph{ogs5} \strong{additional} bloc.
}
