% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_write_tofile}
\alias{ogs5_write_tofile}
\title{ogs5_write_tofile}
\usage{
ogs5_write_tofile(filename = character(), text_output_fct)
}
\arguments{
\item{filename}{Path of file to be written.}

\item{text_output_fct}{Textoutput to be written.}
}
\description{
Write text output to file.
}
