% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_vtu.R
\name{ogs_read_vtu_file_point_data_all}
\alias{ogs_read_vtu_file_point_data_all}
\title{ogs_read_vtu_file_point_data_all}
\usage{
ogs_read_vtu_file_point_data_all(filename)
}
\arguments{
\item{filename}{\emph{charcter} Path to *\strong{.vtu} file.}
}
\value{
\emph{list}
}
\description{
Read all point data array from a *\strong{.vtu} output file.
}
