% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_vtu.R
\name{ogs_read_vtu_file_point_data_array}
\alias{ogs_read_vtu_file_point_data_array}
\title{ogs_read_vtu_file_point_data_array}
\usage{
ogs_read_vtu_file_point_data_array(
  filename = character(),
  array_name = character()
)
}
\arguments{
\item{filename}{Path to *\strong{.vtu} file. \emph{character}}

\item{array_name}{Name of the array in the *\strong{.vtu} file to be read.
\emph{charcter}}
}
\value{
\emph{numeric} vector.
}
\description{
Read specific point data array from single *\strong{.vtu} file.
}
