% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_vtu.R
\name{ogs_read_vtu_files_point_data}
\alias{ogs_read_vtu_files_point_data}
\title{ogs_read_vtu_files_point_data}
\usage{
ogs_read_vtu_files_point_data(filepath, pcs_type, variable_name = NULL)
}
\arguments{
\item{filepath}{\emph{character} Path to directory of **\emph{.vtu} files.}

\item{pcs_type}{\emph{character} Name of the process defined in the \strong{out} sub bloc.}

\item{variable_name}{\emph{character} Name of the array in the *\strong{.vtu} files, optional.}
}
\value{
\emph{list}
}
\description{
Read specific point data arrays from a series of *\strong{.vtu} files
in a specific directory. The series is specified by the \strong{PCS_TYPE}
given in the corresponding \strong{out} sub bloc (ogs5_obj$input$out$...).
}
