% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_run_sim.R
\name{search_ogs5_bin_path}
\alias{search_ogs5_bin_path}
\title{search_ogs5_bin_path}
\usage{
search_ogs5_bin_path(ogs_exe = "ogs", return_ogs5_bin = FALSE)
}
\arguments{
\item{ogs_exe}{\emph{character} Name of the ogs5 executable.}

\item{return_ogs5_bin}{\emph{boolean} If TRUE returns the path to the ogs5
binary instead of the bin folder path.}
}
\value{
\emph{character} of path to ogs5 bin folder or ogs5 binary
}
\description{
Search paths to ogs5 binary or ogs5 bin folder. Includes
'PATH' environment variable. Returns first match of provided 'ogs_exe'.
}
