% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ext_ref.R
\docType{data}
\name{tidyeval-data}
\alias{tidyeval-data}
\alias{.data}
\title{Data pronoun for tidy evaluation}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
.data
}
\description{
See \code{rlang::\link[rlang:tidyeval-data]{.data}} for details.
}
\keyword{datasets}
