% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_bayesOpt.R
\name{to01}
\alias{to01}
\title{Transform to unit interval}
\usage{
to01(par_df, scale_which = NULL, scale_fun = I, unscale_fun = I)
}
\arguments{
\item{par_df}{\emph{tibble} such as explained in \code{\link[=cal_bayesOpt]{cal_bayesOpt()}} with original values}

\item{scale_which}{\emph{character} (optional) that identifies the parameters to be scaled}

\item{scale_fun}{\emph{function} (optional) to scale the values, e.g. \code{log10()}}

\item{unscale_fun}{\emph{function} (optional) inverse of \code{scale_fun}, e.g. `10**x``}
}
\value{
\emph{tibble} with parameter specification and transformed values in the unit interval
}
\description{
Transform to unit interval
}
