% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracer_exp}
\alias{tracer_exp}
\title{Breaktrough curve of a tracer experiment in a horizontal flow wetland."}
\format{
A data frame with 71 rows and 2 variables:
\describe{
\item{time}{time, in days}
\item{tracer_exp}{tracer concentration, in mg/L}
...
}
}
\usage{
tracer_exp
}
\description{
A dataset containing the the tracer break through curve at the outlet of a
horizontal sub-surface flow treatment wetland.
}
\keyword{datasets}
