% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_options_from_config}
\alias{update_options_from_config}
\title{update_options_from_config}
\usage{
update_options_from_config(
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  file = Sys.getenv("R_CONFIG_FILE", "config.yml")
)
}
\arguments{
\item{config}{\emph{character} Name of the configuration environment. Default is:
'default'.}

\item{file}{\emph{character} Name of the configuration file. Default is:
'config.yml'.}
}
\value{
\emph{boolean} TRUE if configuration was update, FALSE if not.
}
\description{
Tries to update package options from a user-defined
configuration file.
}
