% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_additional.R
\name{valid_ogs5_additional}
\alias{valid_ogs5_additional}
\title{valid_ogs5_additional}
\usage{
valid_ogs5_additional(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_additional} to validate.}
}
\description{
Validator for \emph{ogs5_additional} base class
}
