% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_bc.R
\name{valid_ogs5_bc}
\alias{valid_ogs5_bc}
\title{valid_ogs5_bc}
\usage{
valid_ogs5_bc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_bc} to validate.}
}
\description{
Validator for \emph{ogs5_bc} base class
}
