% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_bc.R
\name{valid_ogs5_bc_condition}
\alias{valid_ogs5_bc_condition}
\title{valid_ogs5_bc_condition}
\usage{
valid_ogs5_bc_condition(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_bc_condition} to validate.}
}
\description{
Validator for \emph{ogs5_bc_condition} base class
}
