% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_cct.R
\name{valid_ogs5_cct}
\alias{valid_ogs5_cct}
\title{valid_ogs5_cct}
\usage{
valid_ogs5_cct(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_cct} to validate.}
}
\description{
Validator for \emph{ogs5_cct} base class
}
