% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_fct.R
\name{valid_ogs5_fct}
\alias{valid_ogs5_fct}
\title{valid_ogs5_fct}
\usage{
valid_ogs5_fct(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_fct} to validate.}
}
\description{
Validator for \emph{ogs5_fct} base class
}
