% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gem.R
\name{valid_ogs5_gem}
\alias{valid_ogs5_gem}
\title{valid_ogs5_gem}
\usage{
valid_ogs5_gem(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gem} to validate.}
}
\description{
Validator for \emph{ogs5_gem} base class
}
