% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli}
\alias{valid_ogs5_gli}
\title{valid_ogs5_gli}
\usage{
valid_ogs5_gli(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gli} to validate.}
}
\description{
Validator for \emph{ogs5_gli} base class
}
