% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli_points}
\alias{valid_ogs5_gli_points}
\title{valid_ogs5_gli_points}
\usage{
valid_ogs5_gli_points(x)
}
\arguments{
\item{x}{Object of class c("tbl", "tbl_df", "data.frame").}
}
\description{
Validator for ogs5_gli_points
}
