% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli_polyline}
\alias{valid_ogs5_gli_polyline}
\title{valid_ogs5_gli_polyline}
\usage{
valid_ogs5_gli_polyline(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gli_polyline} to validate.}
}
\description{
Validator for \emph{ogs5_gli_polyline} base class
}
