% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli_polylines}
\alias{valid_ogs5_gli_polylines}
\title{valid_ogs5_gli_polylines}
\usage{
valid_ogs5_gli_polylines(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gli_polylines} to validate.}
}
\description{
Validator for \emph{ogs5_gli_polylines} base class
}
