% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli_surface}
\alias{valid_ogs5_gli_surface}
\title{valid_ogs5_gli_surface}
\usage{
valid_ogs5_gli_surface(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gli_surface} to validate.}
}
\description{
Validator for \emph{ogs5_gli_surface} base class
}
