% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_gli.R
\name{valid_ogs5_gli_surfaces}
\alias{valid_ogs5_gli_surfaces}
\title{valid_ogs5_gli_surfaces}
\usage{
valid_ogs5_gli_surfaces(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_gli_surfaces} to validate.}
}
\description{
Validator for \emph{ogs5_gli_surfaces} base class
}
