% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_ic.R
\name{valid_ogs5_ic}
\alias{valid_ogs5_ic}
\title{valid_ogs5_ic}
\usage{
valid_ogs5_ic(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_ic} to validate.}
}
\description{
Validator for \emph{ogs5_ic} base class
}
