% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_krc.R
\name{valid_ogs5_krc}
\alias{valid_ogs5_krc}
\title{valid_ogs5_krc}
\usage{
valid_ogs5_krc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_krc} to validate.}
}
\description{
Validator for \emph{ogs5_krc} base class
}
