% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_mcp.R
\name{valid_ogs5_mcp}
\alias{valid_ogs5_mcp}
\title{valid_ogs5_mcp}
\usage{
valid_ogs5_mcp(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_mcp} to validate.}
}
\description{
Validator for \emph{ogs5_mcp} base class
}
