% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_mcp.R
\name{valid_ogs5_mcp_component}
\alias{valid_ogs5_mcp_component}
\title{valid_ogs5_mcp_component}
\usage{
valid_ogs5_mcp_component(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_mcp_component} to validate.}
}
\description{
Validator for \emph{ogs5_mcp_component} base class
}
