% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_mfp.R
\name{valid_ogs5_mfp}
\alias{valid_ogs5_mfp}
\title{valid_ogs5_mfp}
\usage{
valid_ogs5_mfp(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_mfp} to validate.}
}
\description{
Validator for \emph{ogs5_mfp} base class
}
