% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_mmp.R
\name{valid_ogs5_mmp}
\alias{valid_ogs5_mmp}
\title{valid_ogs5_mmp}
\usage{
valid_ogs5_mmp(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_mmp} to validate.}
}
\description{
Validator for \emph{ogs5_mmp} base class
}
