% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_msh.R
\name{valid_ogs5_msh}
\alias{valid_ogs5_msh}
\title{valid_ogs5_msh}
\usage{
valid_ogs5_msh(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_msh} to validate.}
}
\description{
Validator for \emph{ogs5_msh} base class
}
