% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_msp.R
\name{valid_ogs5_msp}
\alias{valid_ogs5_msp}
\title{valid_ogs5_msp}
\usage{
valid_ogs5_msp(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_msp} to validate.}
}
\description{
Validator for \emph{ogs5_msp} base class
}
