% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_num.R
\name{valid_ogs5_num}
\alias{valid_ogs5_num}
\title{valid_ogs5_num}
\usage{
valid_ogs5_num(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_num} to validate.}
}
\description{
Validator for \emph{ogs5_num} base class
}
