% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_out.R
\name{valid_ogs5_out}
\alias{valid_ogs5_out}
\title{valid_ogs5_out}
\usage{
valid_ogs5_out(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_out} to validate.}
}
\description{
Validator for \emph{ogs5_out} base class
}
