% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_pcs.R
\name{valid_ogs5_pcs}
\alias{valid_ogs5_pcs}
\title{valid_ogs5_pcs}
\usage{
valid_ogs5_pcs(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_pcs} to validate.}
}
\description{
Validator for \emph{ogs5_pcs} base class
}
