% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_pcs.R
\name{valid_ogs5_pcs_process}
\alias{valid_ogs5_pcs_process}
\title{valid_ogs5_pcs_process}
\usage{
valid_ogs5_pcs_process(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_pcs_process} to validate.}
}
\description{
Validator for \emph{ogs5_pcs_process} base class
}
