% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_pqc.R
\name{valid_ogs5_pqc}
\alias{valid_ogs5_pqc}
\title{valid_ogs5_pqc}
\usage{
valid_ogs5_pqc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_pqc} to validate.}
}
\description{
Validator for \emph{ogs5_pqc} base class
}
