% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_pqc.R
\name{valid_ogs5_pqc_filebloc}
\alias{valid_ogs5_pqc_filebloc}
\title{valid_ogs5_pqc_filebloc}
\usage{
valid_ogs5_pqc_filebloc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_pqc_filebloc} to validate.}
}
\description{
Validator for \emph{ogs5_pqc_filebloc} base class
}
