% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_pqc.R
\name{valid_ogs5_pqc_skeybloc}
\alias{valid_ogs5_pqc_skeybloc}
\title{valid_ogs5_pqc_skeybloc}
\usage{
valid_ogs5_pqc_skeybloc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_pqc_skeybloc} to validate.}
}
\description{
Validator for \emph{ogs5_pqc_skeybloc} base class
}
