% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_rei.R
\name{valid_ogs5_rei}
\alias{valid_ogs5_rei}
\title{valid_ogs5_rei}
\usage{
valid_ogs5_rei(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_rei} to validate.}
}
\description{
Validator for \emph{ogs5_rei} base class
}
