% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs_rfd.R
\name{valid_ogs5_rfd}
\alias{valid_ogs5_rfd}
\title{valid_ogs5_rfd}
\usage{
valid_ogs5_rfd(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_rfd} to validate.}
}
\description{
Validator for \emph{ogs5_rfd} base class
}
