% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_st.R
\name{valid_ogs5_st}
\alias{valid_ogs5_st}
\title{valid_ogs5_st}
\usage{
valid_ogs5_st(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_st} to validate.}
}
\description{
Validator for \emph{ogs5_st} base class
}
