% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_st.R
\name{valid_ogs5_st_condition}
\alias{valid_ogs5_st_condition}
\title{valid_ogs5_st_condition}
\usage{
valid_ogs5_st_condition(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_st_condition} to validate.}
}
\description{
Validator for \emph{ogs5_st_condition} base class
}
