% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_tim.R
\name{valid_ogs5_tim}
\alias{valid_ogs5_tim}
\title{valid_ogs5_tim}
\usage{
valid_ogs5_tim(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_tim} to validate.}
}
\description{
Validator for \emph{ogs5_tim} base class
}
