% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_tim.R
\name{valid_ogs5_tim_bloc}
\alias{valid_ogs5_tim_bloc}
\title{valid_ogs5_tim_bloc}
\usage{
valid_ogs5_tim_bloc(x)
}
\arguments{
\item{x}{Object of class \emph{ogs5_tim_bloc} to validate.}
}
\description{
Validator for \emph{ogs5_tim_bloc} base class
}
