% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{valid_url}
\alias{valid_url}
\title{Check if url is available}
\usage{
valid_url(url_in)
}
\arguments{
\item{url_in}{An url or file path}
}
\value{
\emph{logical} if url can be accessed
}
\description{
Base R solution to check url to avoid dependency RCurl
}
