% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_methods.R
\name{cal_change_parameters}
\alias{cal_change_parameters}
\title{change input parameters of an \emph{ogs5} object}
\usage{
cal_change_parameters(
  ogs5_obj,
  par_df,
  ensemble_path = NULL,
  ensemble_name = NULL
)
}
\arguments{
\item{ogs5_obj}{\emph{ogs5}}

\item{par_df}{\emph{tibble} with parameters such as described in \code{\link{cal_bayesOpt}}}

\item{ensemble_path}{\emph{character} path where ensemble for several parameters
should be written and run.}
}
\value{
An \emph{ogs5} or \emph{ens} object with changed input parameters.
}
\description{
change input parameters of an \emph{ogs5} object
}
