% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_methods.R
\name{cal_create_calibration_set}
\alias{cal_create_calibration_set}
\title{Helper function to create a calibration set for ogs5 input parameters}
\usage{
cal_create_calibration_set(range_list)
}
\arguments{
\item{range_list}{\emph{list} of vectors containing parameter locations in
an \emph{ogs5}-object and its range.}
}
\value{
\emph{data.frame} that can be handed over to \code{\link[=cal_sample_parameters]{cal_sample_parameters()}}
}
\description{
Helper function to create a calibration set for ogs5 input parameters
}
\examples{
\dontrun{
# define range list
range_list <- list(
    c("mmp$MEDIUM_PROPERTIES1$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-4, 1.0e-2),
    c("mmp$MEDIUM_PROPERTIES2$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-9, 1.0e-4),
    c("mmp$MEDIUM_PROPERTIES3$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-7, 1.0e-3),
    c("mmp$MEDIUM_PROPERTIES4$PERMEABILITY_TENSOR", "ISOTROPIC", 1.0e-7, 1.0e-3)
)
# convert into a calibration set
calibration_set <- cal_create_calibration_set(range_list)
calibration_set
}
}
