% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_class.R
\name{create_ens}
\alias{create_ens}
\title{create_ens}
\usage{
create_ens(
  base_sim = list(),
  parameter_tbl = dplyr::tbl(),
  name = character(NULL),
  path = character(NULL)
)
}
\arguments{
\item{base_sim}{Instance of class \emph{ogs5} to be used as basis. \emph{ogs5}}

\item{parameter_tbl}{\emph{tibble()}, \emph{tbl()} or \emph{data.frame()} defining the parameters and values for ensemble
runs.}

\item{name}{\emph{character} Name for the ensemble.}

\item{path}{\emph{character} Path where to place all individual ensemble runs.}
}
\value{
Instance of class \emph{ens}
}
\description{
Constructor for class \emph{ens}. Takes an instance of class \emph{ogs5}
as basis to create a collection of \emph{ogs5} instances to be used in ensemble
simulations. Class \emph{ens} provides wrapper functions to define, execute and
process multiple simulations.
}
\examples{
para_df <- expand.grid(mass_dispersion = seq(0.1, 2.5, length.out = 10),
              tracer_input_concentration = seq(0.1, 1, length.out = 5)) \%>\%
              tibble::as_tibble()

ens1 <- create_ens(base_sim = ex1, parameter_tbl = para_df,
                  name = "tracersim", path = "path/to/my/ens")
}
