% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ens_get_output}
\alias{ens_get_output}
\title{ens_get_output}
\usage{
ens_get_output(x, type = "all", ...)
}
\arguments{
\item{x}{Instanc of class \emph{ens} containing the simulation objects.}

\item{type}{\emph{character} If 'all' retrieve all produced output, if 'specific'
retrieve output stated in a certain \emph{OUT} bloc of the \emph{ogs5} simulation
objects.}

\item{...}{\emph{list(character)} If \emph{type} = 'specific', state names  of the \emph{OUT}
blocs you wish to retrieve output for. Arguments will be passed to
' ogs5_get_output_specific()'.}
}
\value{
Extended \emph{x} with added output.
}
\description{
Wrapper to retrieve output for all included \emph{ogs5} simulation
objects. It calls on the functions \code{\link[=ogs5_get_output_all]{ogs5_get_output_all()}} and
\code{\link[=ogs5_get_output_specific]{ogs5_get_output_specific()}} of the \strong{r2ogs} package.
}
