% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_krc_bloc.R
\name{input_add_krc_bloc}
\alias{input_add_krc_bloc}
\title{input_add_krc_bloc}
\usage{
input_add_krc_bloc(
  x = list(),
  krc_name = character(NULL),
  mkey = character(NULL),
  ACTIVITY_MODEL = NULL,
  ALLOW_REACTIONS = NULL,
  BACTERIACAPACITY = NULL,
  BACTERIAL_YIELD = NULL,
  BACTERIANAME = NULL,
  BACTERIA_SPECIFIC_CAPACITY = NULL,
  BASETERM = NULL,
  CALC_SHERWOOD = NULL,
  CALC_SHERWOOD_MODIFIED = NULL,
  CHEMAPPNAME = NULL,
  COPY_CONCENTRATIONS = NULL,
  D50 = NULL,
  DEBUG_OUTPUT = FALSE,
  DM = NULL,
  DS = NULL,
  EQUATION = NULL,
  EQUILIBRIUM_CONSTANT = NULL,
  EXCHANGE_PARAMETERS = NULL,
  GAS_DISSOLUTION = NULL,
  GEOMETRY = NULL,
  GRAIN_SPHERE_RATIO = NULL,
  GROWTH = NULL,
  INHIBITIONTERMS = NULL,
  INITIAL_TIMESTEP = NULL,
  INTERFACIAL_AREA = NULL,
  ISOTOPE_FRACTIONATION = NULL,
  LAGNEAU_BENCHMARK = FALSE,
  LENGTH = NULL,
  MECHANISMTERM = NULL,
  MICROBENAME = NULL,
  MINERALNAME = NULL,
  MIN_BACTERIACONC = NULL,
  MIN_CONCENTRATION_REPLACE = NULL,
  MIN_TIMESTEP = NULL,
  MONODTERMS = NULL,
  MONOD_REACTION_NAME = NULL,
  NAME = NULL,
  NAPL_CONTENT_INI = NULL,
  NAPL_CONTENT_RES = NULL,
  NAPL_PROPERTIES = NULL,
  NO_REACTIONS = NULL,
  OMEGA_THRESHOLD = NULL,
  PRECIPITATION_BY_BASETERM_ONLY = FALSE,
  PRECIPITATION_EXPONENT = NULL,
  PRECIPITATION_FACTOR = NULL,
  PRODUCTIONSTOCH = NULL,
  PRODUCTIONTERMS = NULL,
  RATECONSTANT = NULL,
  RATE_EXPONENTS = NULL,
  REACTION_DEACTIVATION = NULL,
  REACTION_ORDER = NULL,
  REACTIVE_SURFACE_AREA = NULL,
  RELATIVE_ERROR = NULL,
  SCALE_DCDT = FALSE,
  SHERWOOD_MODEL = NULL,
  SOLVER_TYPE = NULL,
  SORPTION_TYPE = NULL,
  SORT_NODES = FALSE,
  STANDARD_GIBBS_ENERGY = NULL,
  SURFACES = NULL,
  SWITCH_OFF_GEOMETRY = NULL,
  TEMPERATURE_DEPENDENCE = NULL,
  THRESHHOLDTERMS = NULL,
  TORTUOSITY = NULL,
  TYPE = NULL,
  UI = NULL,
  drmc = NULL,
  drmc_PARAMETERS = NULL
)
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{krc_name}{Name of the \strong{krc} sub bloc. \emph{character}}

\item{mkey}{Main keyword of the \emph{ogs5_krc_bloc}: c("BLOB_PROPERTIES",
"KINREACTIONDATA", "MICROBE_PROPERTIES", "REACTION"). \emph{character}}

\item{drmc_PARAMETERS}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{krc} bloc of \emph{ogs5} for defining kinetic
reactions. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/krc.html}{ogs5 keyword docs}.
}
