% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_mcp_bloc.R
\name{input_add_mcp_bloc}
\alias{input_add_mcp_bloc}
\title{input_add_mcp_bloc}
\usage{
input_add_mcp_bloc(
  x = list(),
  ACENTRIC_FACTOR = NULL,
  A_ZERO = NULL,
  BUBBLE_VELOCITY = NULL,
  CRITICAL_PRESSURE = NULL,
  CRITICAL_TEMPERATURE = NULL,
  DECAY = NULL,
  DIFFUSION = NULL,
  FLUID_ID = NULL,
  FLUID_PHASE = NULL,
  FORMULA = NULL,
  ISOTHERM = NULL,
  MAXIMUM_AQUEOUS_SOLUBILITY = NULL,
  MINERAL_DENSITY = NULL,
  MOBILE = NULL,
  MOLAR_DENSITY = NULL,
  MOLAR_VOLUME = NULL,
  MOLAR_WEIGHT = NULL,
  MOL_MASS = NULL,
  NAME = NULL,
  OutputMassOfComponentInModel = NULL,
  TRANSPORT_PHASE = NULL,
  VALENCE = NULL,
  VOLUME_DIFFUSION = NULL
)
}
\arguments{
\item{x}{\emph{ogs5} simulation object.}

\item{NAME}{Component name.}

\item{VOLUME_DIFFUSION}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{mcp} bloc of \emph{ogs5} for defining component
properties. For additional documentatoin of the input parameters
see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/mcp.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
ogs5_obj <- input_add_mcp_bloc(ogs5_obj, NAME = "C(4)",
                               MOBILE = "1",
                               DIFFUSION = "1 0.0e-9",
                               VALENCE = "-2")
}
