% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_mmp_bloc.R
\name{input_add_mmp_bloc}
\alias{input_add_mmp_bloc}
\title{input_add_mmp_bloc}
\usage{
input_add_mmp_bloc(
  x = list(),
  CAPILLARY_PRESSURE = NULL,
  CHANNEL = FALSE,
  COMPOUND_DEPENDENT_DT = NULL,
  CONDUCTIVITY_MODEL = NULL,
  CONVERSION_FACTOR = NULL,
  DATA = NULL,
  DIFFUSION = NULL,
  DIS_TYPE = NULL,
  ELEMENT_VOLUME_MULTIPLYER = NULL,
  EVAPORATION = NULL,
  FLOWLINEARITY = NULL,
  GEOMETRY_AREA = NULL,
  GEOMETRY_DIMENSION = NULL,
  GEOMETRY_INCLINATION = NULL,
  GEO_TYPE = NULL,
  HEAT_DISPERSION = NULL,
  HEAT_TRANSFER = NULL,
  INTERPHASE_FRICTION = NULL,
  MASS_DISPERSION = NULL,
  MMP_TYPE = NULL,
  MSH_TYPE = NULL,
  NAME = NULL,
  ORGANIC_CARBON = NULL,
  PARTICLE_DIAMETER = NULL,
  PCS_TYPE = NULL,
  PERMEABILITY_DISTRIBUTION = NULL,
  PERMEABILITY_FUNCTION_DEFORMATION = NULL,
  PERMEABILITY_FUNCTION_EFFSTRESS = NULL,
  PERMEABILITY_FUNCTION_POROSITY = NULL,
  PERMEABILITY_FUNCTION_PRESSURE = NULL,
  PERMEABILITY_FUNCTION_STRAIN = NULL,
  PERMEABILITY_FUNCTION_STRESS = NULL,
  PERMEABILITY_FUNCTION_VELOCITY = NULL,
  PERMEABILITY_SATURATION = NULL,
  PERMEABILITY_TENSOR = NULL,
  POROSITY = NULL,
  POROSITY_DISTRIBUTION = NULL,
  RILL = NULL,
  SPECIFIC_STORAGE = NULL,
  STORAGE = NULL,
  STORAGE_FUNCTION_EFFSTRESS = NULL,
  SURFACE_FRICTION = NULL,
  TORTUOSITY = NULL,
  TRANSFER_COEFFICIENT = NULL,
  UNCONFINED = NULL,
  VOL_BIO = NULL,
  VOL_MAT = NULL,
  WIDTH = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{CAPILLARY_PRESSURE}{ogs5 \strong{mmp} bloc sub key word.}

\item{CHANNEL}{ogs5 \strong{mmp} bloc sub key word.}

\item{COMPOUND_DEPENDENT_DT}{ogs5 \strong{mmp} bloc sub key word.}

\item{CONDUCTIVITY_MODEL}{ogs5 \strong{mmp} bloc sub key word.}

\item{CONVERSION_FACTOR}{ogs5 \strong{mmp} bloc sub key word.}

\item{DATA}{ogs5 \strong{mmp} bloc sub key word.}

\item{DIFFUSION}{ogs5 \strong{mmp} bloc sub key word.}

\item{DIS_TYPE}{ogs5 \strong{mmp} bloc sub key word.}

\item{ELEMENT_VOLUME_MULTIPLYER}{ogs5 \strong{mmp} bloc sub key word.}

\item{EVAPORATION}{ogs5 \strong{mmp} bloc sub key word.}

\item{FLOWLINEARITY}{ogs5 \strong{mmp} bloc sub key word.}

\item{GEOMETRY_AREA}{ogs5 \strong{mmp} bloc sub key word.}

\item{GEOMETRY_DIMENSION}{ogs5 \strong{mmp} bloc sub key word.}

\item{GEOMETRY_INCLINATION}{ogs5 \strong{mmp} bloc sub key word.}

\item{GEO_TYPE}{ogs5 \strong{mmp} bloc sub key word.}

\item{HEAT_DISPERSION}{ogs5 \strong{mmp} bloc sub key word.}

\item{HEAT_TRANSFER}{ogs5 \strong{mmp} bloc sub key word.}

\item{INTERPHASE_FRICTION}{ogs5 \strong{mmp} bloc sub key word.}

\item{MASS_DISPERSION}{ogs5 \strong{mmp} bloc sub key word.}

\item{MMP_TYPE}{ogs5 \strong{mmp} bloc sub key word.}

\item{MSH_TYPE}{ogs5 \strong{mmp} bloc sub key word.}

\item{NAME}{ogs5 \strong{mmp} bloc sub key word.}

\item{ORGANIC_CARBON}{ogs5 \strong{mmp} bloc sub key word.}

\item{PARTICLE_DIAMETER}{ogs5 \strong{mmp} bloc sub key word.}

\item{PCS_TYPE}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_DISTRIBUTION}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_DEFORMATION}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_EFFSTRESS}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_POROSITY}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_PRESSURE}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_STRAIN}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_STRESS}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_FUNCTION_VELOCITY}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_SATURATION}{ogs5 \strong{mmp} bloc sub key word.}

\item{PERMEABILITY_TENSOR}{ogs5 \strong{mmp} bloc sub key word.}

\item{POROSITY}{ogs5 \strong{mmp} bloc sub key word.}

\item{POROSITY_DISTRIBUTION}{ogs5 \strong{mmp} bloc sub key word.}

\item{RILL}{ogs5 \strong{mmp} bloc sub key word.}

\item{SPECIFIC_STORAGE}{ogs5 \strong{mmp} bloc sub key word.}

\item{STORAGE}{ogs5 \strong{mmp} bloc sub key word.}

\item{STORAGE_FUNCTION_EFFSTRESS}{ogs5 \strong{mmp} bloc sub key word.}

\item{SURFACE_FRICTION}{ogs5 \strong{mmp} bloc sub key word.}

\item{TORTUOSITY}{ogs5 \strong{mmp} bloc sub key word.}

\item{TRANSFER_COEFFICIENT}{ogs5 \strong{mmp} bloc sub key word.}

\item{UNCONFINED}{ogs5 \strong{mmp} bloc sub key word.}

\item{VOL_BIO}{ogs5 \strong{mmp} bloc sub key word.}

\item{VOL_MAT}{ogs5 \strong{mmp} bloc sub key word.}

\item{WIDTH}{ogs5 \strong{mmp} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{mmp} bloc of \emph{ogs5} for defining material
properties. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/mmp.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
ogs5_obj <- input_add_mmp_bloc(ogs5_obj, NAME = "MEDIUM_PROPERTIES1",
                               GEOMETRY_DIMENSION = "1",
                               GEOMETRY_AREA = "1.000000e+000",
                               POROSITY = "1 0.32",
                               VOL_MAT = "1 0.68",
                               TORTUOSITY = "1 1.000000e+000",
                               PERMEABILITY_TENSOR = "ISOTROPIC 1.15700e-12",
                               MASS_DISPERSION = "1 0.0067 0.1000")
}
