% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_msp_bloc.R
\name{input_add_msp_bloc}
\alias{input_add_msp_bloc}
\title{input_add_msp_bloc}
\usage{
input_add_msp_bloc(
  x = list(),
  BIOT_CONSTANT = NULL,
  CREEP = NULL,
  DENSITY = character(NULL),
  ELASTICITY = NULL,
  EXCAVATION = NULL,
  E_Function = NULL,
  GRAVITY_CONSTANT = NULL,
  MICRO_STRUCTURE_PLAS = NULL,
  NAME = character(NULL),
  NON_REACTIVE_FRACTION = NULL,
  PLASTICITY = NULL,
  REACTIVE_SYSTEM = NULL,
  SOLID_BULK_MODULUS = NULL,
  SPECIFIC_HEAT_SOURCE = NULL,
  STRESS_INTEGRATION_TOLERANCE = NULL,
  STRESS_UNIT = NULL,
  SWELLING_PRESSURE_TYPE = NULL,
  THERMAL = NULL,
  THRESHOLD_DEV_STR = NULL,
  TIME_DEPENDENT_YOUNGS_POISSON = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{NAME}{Name of the solid. \emph{character}}

\item{TIME_DEPENDENT_YOUNGS_POISSON}{}
}
\value{

}
\description{
Adds a sub-bloc to \strong{msp} bloc of \emph{ogs5} for defining solid
properties. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/msp.html}{ogs5 keyword docs}
}
\examples{
ogs5_obj <- input_add_msp_bloc(ogs5_obj, NAME = "SOLID_PROPERTIES1",
                               DENSITY = "1 1.80000e+003")
}
