% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_num_bloc.R
\name{input_add_num_bloc}
\alias{input_add_num_bloc}
\title{input_add_num_bloc}
\usage{
input_add_num_bloc(
  x = list(),
  num_name = character(NULL),
  COUPLED_PROCESS = NULL,
  COUPLING_CONTROL = NULL,
  COUPLING_ITERATIONS = NULL,
  DYNAMIC_DAMPING = NULL,
  ELE_GAUSS_POINTS = NULL,
  ELE_MASS_LUMPING = NULL,
  ELE_SUPG = NULL,
  ELE_UPWINDING = NULL,
  EXTERNAL_SOLVER_OPTION = NULL,
  FEM_FCT = NULL,
  GRAVITY_PROFILE = NULL,
  LINEAR_SOLVER = character(NULL),
  LOCAL_PICARD1 = NULL,
  NON_LINEAR_ITERATION = NULL,
  NON_LINEAR_SOLVER = NULL,
  NON_LINEAR_UPDATE_VELOCITY = NULL,
  OVERALL_COUPLING = NULL,
  PCS_TYPE = NUcharacter(NULL),
  PLASTICITY_TOLERANCE = NULL,
  RENUMBER = NULL,
  TIME_STEPS = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{num_name}{Name of the sub-bloc.}

\item{TIME_STEPS}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{num} bloc of \emph{ogs5} for defining numerical
solver properties. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/num.html}{ogs5 keyword docs}.
}
\examples{
ogs5_obj <- input_add_num_bloc(ogs5_obj, num_name = "NUMERICS1",
                               PCS_TYPE = "GROUNDWATER_FLOW",
                               ELE_GAUSS_POINTS = "3",
                               LINEAR_SOLVER = "2 6 1.e-014 1000 1.0 1 2")
}
