% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_pcs_bloc.R
\name{input_add_pcs_bloc}
\alias{input_add_pcs_bloc}
\title{input_add_pcs_bloc}
\usage{
input_add_pcs_bloc(
  x = list(),
  pcs_name = character(NULL),
  APP_TYPE = NULL,
  BOUNDARY_CONDITION_OUTPUT = FALSE,
  COUNT = NULL,
  CPL_TYPE = NULL,
  DEACTIVATED_SUBDOMAIN = NULL,
  DISSOLVED_CO2_INGAS_PCS_NAME = NULL,
  DISSOLVED_CO2_PCS_NAME = NULL,
  ELEMENT_MATRIX_OUTPUT = NULL,
  GEO_TYPE = NULL,
  MEDIUM_TYPE = NULL,
  MEMORY_TYPE = NULL,
  MSH_TYPE = NULL,
  NEGLECT_H_INI_EFFECT = NULL,
  NUM_TYPE = NULL,
  OutputMassOfGasInModel = FALSE,
  PCS_TYPE = character(NULL),
  PHASE_TRANSITION = NULL,
  PRIMARY_VARIABLE = NULL,
  PROCESSED_BC = NULL,
  RELOAD = NULL,
  SATURATION_SWITCH = FALSE,
  SAVE_ECLIPSE_DATA_FILES = FALSE,
  SIMULATOR = NULL,
  SIMULATOR_MODEL_PATH = NULL,
  SIMULATOR_PATH = NULL,
  SIMULATOR_WELL_PATH = NULL,
  ST_RHS = NULL,
  TIME_CONTROLLED_EXCAVATION = NULL,
  TIM_TYPE = NULL,
  UPDATE_INI_STATE = NULL,
  USE_PRECALCULATED_FILES = FALSE,
  USE_VELOCITIES_FOR_TRANSPORT = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{pcs_name}{Name of the \strong{pcs} sub-bloc.}

\item{USE_VELOCITIES_FOR_TRANSPORT}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{pcs} bloc of \emph{ogs5} for defining processes
For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/pcs.html}{ogs5 keyword docs}.
}
\examples{
ogs5_obj <- input_add_pcs_bloc(ogs5_obj, pcs_name = "PROCESS1",
                               PCS_TYPE = "GROUNDWATER_FLOW",
                               NUM_TYPE = "NEW",
                               ELEMENT_MATRIX_OUTPUT = "0")
}
