% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_st_bloc.R
\name{input_add_st_bloc}
\alias{input_add_st_bloc}
\title{input_add_st_bloc}
\usage{
input_add_st_bloc(
  x = list(),
  st_name = NULL,
  AIR_BREAKING = NULL,
  CHANNEL = NULL,
  COMP_NAME = NULL,
  CONSTRAINED = NULL,
  DISTRIBUTE_VOLUME_FLUX = NULL,
  DIS_TYPE = NULL,
  EXPLICIT_SURFACE_WATER_PRESSURE = NULL,
  FCT_TYPE = NULL,
  GEO_TYPE = NULL,
  MSH_TYPE = NULL,
  NEGLECT_SURFACE_WATER_PRESSURE = NULL,
  NODE_AVERAGING = NULL,
  PCS_TYPE = NULL,
  PRIMARY_VARIABLE = NULL,
  TIME_INTERPOLATION = NULL,
  TIM_TYPE = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{st_name}{Name of the source term.}

\item{TIM_TYPE}{}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{st} bloc of \emph{ogs5} for defining a source
term. For additional documentatoin of the input parameters see
the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/st.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
All arguments except \strong{x} have to be of class \emph{character}.
}
\examples{
ogs5_obj <- input_add_st_bloc(ogs5_obj, st_name = "SOURCE_TERM1",
                              PCS_TYPE = "GROUNDWATER_FLOW",
                              PRIMARY_VARIABLE = "HEAD",
                              GEO_TYPE = "POINT POINT1",
                              DIS_TYPE = "CONSTANT -2.9976852e-006")
}
