% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_add_tim_bloc.R
\name{input_add_tim_bloc}
\alias{input_add_tim_bloc}
\title{input_add_tim_bloc}
\usage{
input_add_tim_bloc(
  x = list(),
  tim_name = NULL,
  CRITICAL_TIME = NULL,
  INDEPENDENT = NULL,
  PCS_TYPE = NULL,
  SUBSTEPS = NULL,
  TIME_END = NULL,
  TIME_START = NULL,
  TIME_STEPS = NULL,
  TIME_CONTROL = NULL,
  TIME_FIXED_POINTS = NULL,
  TIME_SPLITS = NULL,
  TIME_UNIT = NULL
)
}
\arguments{
\item{x}{Simulation object of class \emph{ogs5}.}

\item{tim_name}{Name of the \strong{tim} sub-bloc.}

\item{CRITICAL_TIME}{ogs5 \strong{tim} bloc sub key word.}

\item{INDEPENDENT}{ogs5 \strong{tim} bloc sub key word.}

\item{PCS_TYPE}{ogs5 \strong{tim} bloc sub key word.}

\item{SUBSTEPS}{ogs5 \strong{tim} bloc sub key word.}

\item{TIME_END}{\emph{numeric}}

\item{TIME_START}{\emph{numeric}}

\item{TIME_STEPS}{ogs5 \strong{tim} bloc sub key word.}

\item{TIME_CONTROL}{ogs5 \strong{tim} bloc sub key word.}

\item{TIME_FIXED_POINTS}{ogs5 \strong{tim} bloc sub key word.}

\item{TIME_SPLITS}{ogs5 \strong{tim} bloc sub key word.}

\item{TIME_UNIT}{ogs5 \strong{tim} bloc sub key word.}
}
\value{
Updated \emph{ogs5} object.
}
\description{
Adds a sub-bloc to \strong{tim} bloc of \emph{ogs5} for defining numerical
time discritization. For additional documentatoin of the input parameters
see the \href{https://ogs5-keywords.netlify.app/ogs/wiki/public/doc-auto/by_ext/tim.html}{ogs5 keyword docs}
or have a look at the input scripts from the \href{https://gitlab.opengeosys.org/ag-hydinf/boog-group/r2ogs5-benchmarks}{r2ogs5-benchmarks repository}.
Most arguments have to be of class \emph{character}.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))

ogs5_obj <- input_add_tim_bloc(ogs5_obj, tim_name = "TIME_STEPPING1",
                              PCS_TYPE = "GROUNDWATER_FLOW",
                              TIME_STEPS = "210 100.0",
                              TIME_END = "21000.0",
                              TIME_START = "0.0")
}
