% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_many_tecplots}
\alias{ogs5_read_many_tecplots}
\title{ogs5_read_many_tecplots}
\usage{
ogs5_read_many_tecplots(filepath = character(), geo_object = character())
}
\arguments{
\item{filepath}{Path to *\strong{.tec} files. \emph{character}}

\item{geo_object}{Associated geometry type: c("domain", "POINT", "POLYLINE",
"SURFACE").}
}
\value{
Dataframe.
}
\description{
Reads many *.tec files from one filepath. Wrapper for
\code{\link[=ogs5_read_tecplot]{ogs5_read_tecplot()}}.
}
\examples{
tec_df <- ogs5_read_many_tecplots(filepath = "examples/tmp/ex1",
                                  geo_object = "domain")
}
