% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_tecplot_domain}
\alias{ogs5_read_tecplot_domain}
\title{ogs5_read_tecplot_domain}
\usage{
ogs5_read_tecplot_domain(filename)
}
\arguments{
\item{filename}{Path to *\strong{.tec} file. \emph{character}}
}
\value{
Dataframe.
}
\description{
Read data from *\strong{.tec} file of entire model domain.
}
