% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_read_output_tec.R
\name{ogs5_read_tecplot_polyline}
\alias{ogs5_read_tecplot_polyline}
\title{ogs5_read_tecplot_polyline}
\usage{
ogs5_read_tecplot_polyline(filepath)
}
\arguments{
\item{filepath}{path to the filename}
}
\value{

}
\description{
ogs5_read_tecplot_polyline
}
