% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_run_sim.R
\name{ogs5_run}
\alias{ogs5_run}
\title{ogs5_run}
\usage{
ogs5_run(
  ogs5_obj = list(),
  ogs_exe = NULL,
  run_path = NULL,
  log_output = TRUE,
  log_path = NULL,
  use_mpi = FALSE,
  number_of_cores = NULL,
  wait = TRUE
)
}
\arguments{
\item{ogs5_obj}{\emph{ogs5} simulation object.}

\item{ogs_exe}{\emph{character} Path to \strong{ogs5} executable.}

\item{run_path}{\emph{character} Path to directory where the simulation should be executed.
Default: use 'attributes(ogs5_obj)$sim_path'}

\item{log_output}{\emph{logical} Should a log file be written?}

\item{log_path}{\emph{character} Path of directory to write log file.}

\item{use_mpi}{\emph{logical} Execute parallelized simulation using MPI.}

\item{number_of_cores}{\emph{integer} Use number of computing cores if 'use_mpi=TRUE'.}

\item{wait}{\emph{logical} if 'TRUE', Freeze R-session until simulation is finished.}
}
\description{
Run \strong{ogs5} simulation via call to OS.
}
\examples{
\dontrun{
 ogs5_run(ogs5_obj = sim1, ogs_exe = "inst/ogs/ogs_5.76",
          run_path = NULL,
          log_output = TRUE,
          log_path = "examples/tmp/ex1/log")
}
}
