% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogs5_write_inputfiles.R
\name{ogs5_write_inputfiles}
\alias{ogs5_write_inputfiles}
\title{ogs5_write_inputfiles}
\usage{
ogs5_write_inputfiles(ogs5_obj = list(), type = "all", folderpath = NULL)
}
\arguments{
\item{ogs5_obj}{\emph{ogs5} simulation object}

\item{type}{\emph{character} Type of the \emph{ogs5} main bloc to be written. c("all", "bc", "cct",
"fct", "gem", "gli", "ic", "krc", "mcp", "mfp", "mmp", "msh", "msp", num",
"out", "pcs", "pqc", "rei", "st", "tim", "rfd", "additional")}

\item{folderpath}{\emph{character} Path to write the outpufile(s) to.}
}
\description{
Export \emph{ogs5} object and write ogs5 input files.
}
\examples{
tmp <- tempdir()
ogs5_obj <- create_ogs5(sim_name = "ex1", sim_id = 1L,
                        sim_path = paste0(tmp, "/ex1"))
ogs5_obj <- input_add_pcs_bloc(ogs5_obj, pcs_name = "PROCESS1",
                               PCS_TYPE = "GROUNDWATER_FLOW",
                               NUM_TYPE = "NEW",
                               ELEMENT_MATRIX_OUTPUT = "0")
ogs5_obj <- input_add_bc_bloc(ogs5_obj, bc_name = "BOUNDARY_CONDITION1",
                              PCS_TYPE = "GROUNDWATER_FLOW",
                              PRIMARY_VARIABLE = "HEAD",
                              GEO_TYPE = "POINT POINT0",
                              DIS_TYPE = "CONSTANT 1.0e5")

ogs5_write_inputfiles(ogs5_obj, "all")
}
